refactor(cli): split commands.rs into commands/ submodules
commands.rs had grown to 2918 SLOC (3.6x the 800-line self-scan
limit), holding the run() dispatcher, all 16 per-subcommand runners,
and the entire check pipeline in one file. Extract them into a
commands/ subdirectory keyed by subcommand area, leaving commands.rs
as a thin dispatcher:

  - commands/check.rs (+ check/ for the pipeline's stages:
    effective_config, thresholds, outcome, remediation, footer)
  - commands/analyze.rs  metrics / ops / structured output
  - commands/scan.rs     find / count / dump / functions / list-metrics
  - commands/report.rs   report rendering entry
  - commands/preproc.rs  strip-comments / preproc
  - commands/init.rs     init / baseline scaffolding
  - commands/diff_cmd.rs diff / diff-baseline (named diff_cmd to avoid
                         clashing with the crate::diff git-plumbing module)
  - commands/exemptions.rs

Pure code movement: bodies are verbatim, no logic change. run() and
its dispatcher helpers (with_manifest, load_preproc, the clap-parse
scaffolding) stay in commands.rs; the in-file test module is unchanged.
Each submodule uses `use super::*` and commands.rs re-exports every
helper via `pub(crate) use <area>::*`, so all call sites resolve
unchanged. Helpers that were file-private are widened to pub(crate)
where the dispatcher, a sibling submodule, or commands_tests.rs reaches
them across the new boundary. The published surface (run) is identical.

Baseline refreshed: the stale commands.rs entries (including the
loc.sloc=2918 file-level entry) are re-anchored to their new paths.

commands.rs 2918 -> 301 SLOC; every new file is under the 800 limit
(largest: check.rs at 613).
