+++ big-code-analysis-cli/src/commands.rs
@@ -1,13 +1,18 @@
 //! Top-level command dispatch for the `bca` CLI.
 //!
 //! Owns the public `run()` entry point (called by `bca`'s `main` and
-//! by `xtask` for man-page rendering), the per-command helpers
-//! (`run_command_*`), the `check` subcommand pipeline (`run_check` +
-//! its stage helpers), and the per-file footer rendering used by
-//! `bca check`'s stderr output.
+//! by `xtask` for man-page rendering) and the dispatch scaffolding it
+//! needs: manifest discovery, preproc loading, and the clap-parse /
+//! legacy-hint glue.
 //!
-//! All other CLI plumbing — argument types, the parallel walker, the
-//! legacy-hint scaffolding — lives in `lib.rs` / sibling submodules.
+//! The per-command handlers live in area submodules — `analyze`
+//! (metrics / ops), `scan` (find / count / dump / functions /
+//! list-metrics), `report`, `preproc`, `init`, `diff_cmd`,
+//! `exemptions`, and `check` (whose multi-stage pipeline is further
+//! split under `check/`). Each is re-exported here so `run()` reaches
+//! every handler by its bare name. Argument types, the parallel
+//! walker, and the diagnostic helpers live in `lib.rs` / sibling
+//! crate modules.
 
 use std::collections::{BTreeMap, HashMap};
 use std::io::Write;
@@ -52,1382 +57,18 @@ use crate::{
     write_output_or_stdout, write_stdout_or_die,
 };
 
-fn run_check(
-    globals: GlobalOpts,
-    mut args: CheckArgs,
-    manifest: Option<&Manifest>,
-    preproc: Option<Arc<PreprocResults>>,
-) {
-    // bca: suppress(abc)
-    // Linear check-pipeline orchestration: each stage (threshold resolve,
-    // walk, baseline filter, classify, emit) is already its own helper; the
-    // ABC count is the call/assignment density of wiring them together, not
-    // branching logic.
-    // Merge the check-only manifest keys (baseline / headroom) under the
-    // CLI flags, and take the `[thresholds]` table (hard + soft layers)
-    // as the base for the resolver. `--config` merges on top of it;
-    // `--threshold` overrides win last (see
-    // `validate_and_build_thresholds`).
-    let base_thresholds = match manifest {
-        Some(m) => {
-            m.merge_check(&mut args);
-            m.thresholds()
-        }
-        None => ParsedThresholds::default(),
-    };
-    // Resolve the deprecated `--headroom` / `--strict-exit-codes` aliases
-    // once (issues #688/#666). The manifest merge above has already
-    // folded `[check] exit_codes` into `args.exit_codes` when the CLI
-    // left it unset, so `resolved_exit_codes` reflects both sources with
-    // the CLI value winning in either direction.
-    let tier = args.resolved_tier();
-    let tiered_exit_codes = args.resolved_exit_codes() == Some(crate::ExitCodes::Tiered);
-    let ResolvedThresholds {
-        set,
-        hard_limits,
-        provenance,
-    } = validate_and_build_thresholds(&mut args, base_thresholds, tier);
-    // `--print-effective-config` is a read-only debug aid: print the
-    // resolved configuration and exit 0 before the walk. clap already
-    // rejects pairing with `--write-baseline` (conflicts_with), so by
-    // the time we get here the flag is unambiguous.
-    if let Some(format) = args.print_effective_config {
-        print_effective_config(
-            &globals,
-            &args,
-            &set,
-            manifest,
-            format,
-            tier,
-            tiered_exit_codes,
-        );
-        return;
-    }
-    let scope = resolve_diff_scope(&args);
-    // Clone globals for the remediation builder: `run_check_walk`
-    // consumes `globals` by value (it passes through to `run_walk`
-    // which spawns worker threads with ownership), but
-    // `format_remediation_block` needs the resolved `--paths` /
-    // `--exclude` set to compose a copy-paste-safe refresh command.
-    let globals_for_remediation = globals.clone();
-    let (violations, files_dispatched) = run_check_walk(globals, &args, preproc, set);
+mod analyze;
+mod check;
+mod diff_cmd;
+mod exemptions;
+mod init;
+mod preproc;
+mod report;
+mod scan;
 
-    if files_dispatched.load(Ordering::Relaxed) == 0 {
-        // No files survived `--paths` expansion + `--include`/`--exclude`
-        // filtering. Treat this as a tool error (exit 1), not a clean
-        // pass (exit 0): a typo in `--paths` would otherwise silently
-        // green-light CI.
-        die("bca check: no input files matched; check --paths, --include, --exclude");
-    }
-
-    // Drop offenders from `[check.exclude]` files (#378) before *any*
-    // downstream consumer sees them — so `--write-baseline` never
-    // records the structural exemptions and the gate never fails on
-    // them. Applied after the empty-input guard above: exempt files are
-    // still walked and counted, only their violations are dropped.
-    let violations = apply_check_exclude(
-        violations,
-        &args,
-        &globals_for_remediation.paths,
-        globals_for_remediation.paths_from.as_deref(),
-    );
-
-    // `--write-baseline <path>` writes there; a bare `--write-baseline`
-    // is resolved to the manifest `baseline` by `merge_check` (#496), so
-    // a remaining `Some(None)` means no path was given and no manifest
-    // `baseline` exists — a hard error rather than a silent no-write.
-    if let Some(target) = args.write_baseline.as_ref() {
-        let path = target.clone().unwrap_or_else(|| {
-            die(
-                "--write-baseline needs a path: pass one (`--write-baseline <file>`) \
-                 or set a `baseline` key in bca.toml",
-            )
-        });
-        write_check_baseline(violations, &path, provenance);
-        return;
-    }
-
-    let pairs = filter_by_baseline(
-        violations,
-        args.baseline.as_deref(),
-        args.baseline_line_tolerance
-            .unwrap_or(baseline::DEFAULT_LINE_TOLERANCE),
-        args.baseline_fuzzy_match.unwrap_or(false),
-        provenance,
-        args.report_suppressed,
-    );
-    // Apply `--changed-only` diff-scope filtering to ALL offenders before
-    // splitting, so the suppressed set surfaced in the report respects the
-    // touched-file scope exactly as the active set does — otherwise
-    // `--changed-only --report-suppressed` would leak suppressed debt from
-    // files outside the diff. With `--report-suppressed` off this is the
-    // original pre-feature ordering (filter, then everything is active).
-    let pairs = apply_changed_only(pairs, scope.as_ref(), args.changed_only);
-    // Split the report-only suppressed debt — in-source markers
-    // (`v.suppressed`) plus baseline-covered offenders
-    // (`Coverage::Covered`), present only under `--report-suppressed` — from
-    // the active offenders. Suppressed debt is surfaced in the code-scan
-    // document but never reaches the gate: exit code, stderr stream, and
-    // remediation are all driven by `active` alone. The default path leaves
-    // `suppressed` empty, so behaviour is byte-for-byte unchanged.
-    let (suppressed, active): (Vec<_>, Vec<_>) = pairs.into_iter().partition(|(v, coverage)| {
-        v.suppressed || matches!(coverage, Some(Coverage::Covered { .. }))
-    });
-    let any_violations = !active.is_empty();
-    // Categorise the active violations for the exit-code contract (#385)
-    // before `emit_check_results` consumes them.
-    let outcome = classify_check_outcome(&active, tier.tier(), &hard_limits);
-    // Build the remediation block ONLY when we have something to
-    // remediate. Empty active set (clean run) gets no trailing block —
-    // there is no baseline to refresh and no artifact worth pointing
-    // at. Suppressed debt is informational and never remediated here.
-    let remediation = if any_violations {
-        format_remediation_block(&globals_for_remediation, &args)
-    } else {
-        None
-    };
-    emit_check_results(
-        active,
-        suppressed,
-        &args,
-        scope.as_ref(),
-        remediation.as_deref(),
-    );
-
-    // `--no-fail` always forces exit 0; otherwise map the outcome to the
-    // process exit code (tiered when `--strict-exit-codes` is set, the
-    // stable 0/1/2 contract otherwise). A clean run returns `None` and
-    // the process exits 0 implicitly.
-    if !args.no_fail
-        && let Some(code) = outcome.exit_code(tiered_exit_codes)
-    {
-        process::exit(code);
-    }
-}
-
-/// Serialize the resolved threshold/check configuration to stdout.
-/// Used by `--print-effective-config` to surface the post-merge view
-/// of every layer (`--config` TOML + repeated `--threshold` CLI
-/// overrides) without running the check.
-///
-/// The output shape is intentionally a strict superset of what
-/// `--config` consumes: piping the TOML form back through `--config`
-/// reproduces the same `ThresholdSet`. JSON is offered for tooling
-/// pipelines (CI dashboards, IDE plugins) that prefer structured data
-/// over TOML — the same field names; same shape.
-///
-/// The resolved layers (headroom scaling per #373, `[thresholds.soft]`
-/// / `--tier` per #375, the tiered exit-code style per #385) are already
-/// folded into the serialized view; future layers (baseline state per
-/// #381) will extend `EffectiveConfig` additively. This printer is the
-/// single place that needs to learn about them.
-fn print_effective_config(
-    globals: &GlobalOpts,
-    args: &CheckArgs,
-    set: &ThresholdSet,
-    manifest: Option<&Manifest>,
-    format: PrintConfigFormat,
-    tier: TierSpec,
-    tiered_exit_codes: bool,
-) {
-    let effective =
-        EffectiveConfig::from_resolved(globals, args, set, manifest, tier, tiered_exit_codes);
-    let serialized = match format {
-        PrintConfigFormat::Toml => toml::to_string_pretty(&effective)
-            .unwrap_or_else(|e| die(format_args!("serialize effective config to TOML: {e}"))),
-        PrintConfigFormat::Json => serde_json::to_string_pretty(&effective)
-            .unwrap_or_else(|e| die(format_args!("serialize effective config to JSON: {e}"))),
-    };
-    write_stdout_or_die(serialized.as_bytes());
-    // TOML's `to_string_pretty` already ends with a newline; JSON's
-    // `to_string_pretty` does not. Normalize so consumers piping into
-    // `--config` or `jq` see a clean trailing newline either way.
-    if !serialized.ends_with('\n') {
-        write_stdout_or_die(b"\n");
-    }
-}
-
-/// Resolved view of `bca check` configuration after layer merge.
-///
-/// Mirrors the [`ThresholdConfig`][crate::thresholds::ThresholdConfig]
-/// schema for the `[thresholds]` table so the TOML form is directly
-/// consumable via `--config`. The `[check]` table reports the
-/// filtering/scoping inputs (paths, include/exclude globs, suppression
-/// policy, etc.) that affect which functions are even considered for
-/// threshold comparison; those fields are informational and ignored
-/// by `--config`.
-#[derive(serde::Serialize)]
-struct EffectiveConfig {
-    thresholds: BTreeMap<String, f64>,
-    check: EffectiveCheck,
-}
-
-#[derive(serde::Serialize)]
-struct EffectiveCheck {
-    paths: Vec<String>,
-    include: Vec<String>,
-    exclude: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    exclude_from: Option<String>,
-    /// Resolved `[check.exclude]` globs (#378): files analysed and
-    /// reported but exempt from the gate. Empty when unset.
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    check_exclude: Vec<String>,
-    /// Source file for additional `check_exclude` globs
-    /// (`--check-exclude-from` / `[check] exclude_from`), if any.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    check_exclude_from: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    paths_from: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    baseline: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    config: Option<String>,
-    /// Path to the auto-discovered `bca.toml` whose keys were merged
-    /// under the CLI flags, if any. Provenance for the resolved view:
-    /// signals that values not traceable to a flag came from here.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    manifest: Option<String>,
-    no_fail: bool,
-    no_suppress: bool,
-    /// Whether `--report-suppressed` is active: suppressed/baselined debt
-    /// is kept in the SARIF offender document (as `suppressions` entries)
-    /// instead of dropped. Surfaced here because it changes the emitted
-    /// offender set — a gate-relevant input the resolved view previously
-    /// omitted (#704).
-    report_suppressed: bool,
-    no_ignore: bool,
-    no_skip_generated: bool,
-    exclude_tests: bool,
-    changed_only: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    since: Option<String>,
-    /// The soft-tier scale ratio applied to the config-derived limits,
-    /// if any (the `RATIO` in `--tier=soft=RATIO`, issue #688). Recorded
-    /// for provenance: the `[thresholds]` table above already shows the
-    /// post-scaling values, so this is the one signal that distinguishes
-    /// "limit 14.25 because config said 15 × 0.95" from "limit 14.25
-    /// because config literally said 14.25".
-    #[serde(skip_serializing_if = "Option::is_none")]
-    headroom: Option<f64>,
-    /// Which tier the `thresholds` table above was resolved for
-    /// (`"hard"` or `"soft"`, issue #375). The limits shown already
-    /// reflect any `[thresholds.soft]` merge or soft-ratio scaling, so
-    /// this field is the one signal that records *which* tier produced
-    /// them.
-    tier: &'static str,
-    /// Which exit-code contract is in force (#385/#666): `"default"`
-    /// (the stable 0/1/2 codes) or `"tiered"` (the 2-5 severity split,
-    /// enabled by `--exit-codes=tiered` or `[check] exit_codes`).
-    exit_codes: &'static str,
-    /// The `--baseline-line-tolerance` override, if set (issue #377).
-    /// Absent means the built-in default applies.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    baseline_line_tolerance: Option<usize>,
-    /// Whether the body-hash fuzzy fallback (`--baseline-fuzzy-match`)
-    /// is active. Only meaningful alongside `baseline`.
-    baseline_fuzzy_match: bool,
-}
-
-impl EffectiveConfig {
-    /// Project the resolved `ThresholdSet` + the original CLI args into
-    /// a serializable view. Paths are rendered with [`Path::display`]
-    /// because the printed config is informational; `--config` only
-    /// reads the `[thresholds]` table back, where keys/values are pure
-    /// ASCII metric names + numbers and round-trip exactly.
-    fn from_resolved(
-        globals: &GlobalOpts,
-        args: &CheckArgs,
-        set: &ThresholdSet,
-        manifest: Option<&Manifest>,
-        tier: TierSpec,
-        tiered_exit_codes: bool,
-    ) -> Self {
-        let thresholds: BTreeMap<String, f64> = set
-            .iter()
-            .map(|(name, limit)| (name.to_owned(), limit))
-            .collect();
-        let check = EffectiveCheck {
-            paths: globals
-                .paths
-                .iter()
-                .map(|p| p.display().to_string())
-                .collect(),
-            include: globals.include.clone(),
-            exclude: globals.exclude.clone(),
-            exclude_from: globals
-                .exclude_from
-                .as_ref()
-                .map(|p| p.display().to_string()),
-            check_exclude: args.check_exclude.clone(),
-            check_exclude_from: args
-                .check_exclude_from
-                .as_ref()
-                .map(|p| p.display().to_string()),
-            paths_from: globals.paths_from.as_ref().map(|p| p.display().to_string()),
-            baseline: args.baseline.as_ref().map(|p| p.display().to_string()),
-            config: args.config.as_ref().map(|p| p.display().to_string()),
-            manifest: manifest.map(|m| m.path().display().to_string()),
-            no_fail: args.no_fail,
-            no_suppress: args.no_suppress,
-            report_suppressed: args.report_suppressed,
-            no_ignore: globals.no_ignore,
-            no_skip_generated: globals.no_skip_generated,
-            exclude_tests: globals.exclude_tests,
-            changed_only: args.changed_only,
-            since: args.since.clone(),
-            headroom: tier.ratio(),
-            tier: tier.tier().as_str(),
-            exit_codes: if tiered_exit_codes {
-                "tiered"
-            } else {
-                "default"
-            },
-            baseline_line_tolerance: args.baseline_line_tolerance,
-            baseline_fuzzy_match: args.baseline_fuzzy_match.unwrap_or(false),
-        };
-        Self { thresholds, check }
-    }
-}
-
-/// Default soft-tier scale applied when `--tier=soft` is requested with
-/// neither a `[thresholds.soft]` table nor an explicit `--headroom`. A
-/// concrete default keeps `--tier=soft` from being a silent no-op (the
-/// "config error" the issue #375 resolution order warns against) — it
-/// always produces a band tighter than the hard gate.
-const DEFAULT_SOFT_HEADROOM: f64 = 0.95;
-
-/// Resolved threshold layers handed back to [`run_check`].
-///
-/// `set` is the gate the walker compares against (the requested tier's
-/// limits). `hard_limits` is the hard-tier limit per metric — equal to
-/// `set` at the hard tier, but the *un-scaled* ceilings at the soft
-/// tier, so [`classify_check_outcome`] can tell a soft-band
-/// encroachment apart from a true hard breach (#385).
-struct ResolvedThresholds {
-    set: Arc<ThresholdSet>,
-    hard_limits: BTreeMap<String, f64>,
-    /// Tier/headroom the gate resolved to (issue #486). Stamped into the
-    /// baseline on `--write-baseline` and compared against a loaded
-    /// baseline's recorded provenance to warn on a stricter-than-baseline
-    /// desync.
-    provenance: baseline::Provenance,
-}
-
-/// Reduce the resolved tier + soft-table presence to the
-/// [`baseline::Provenance`] stamped on a write and compared on a read
-/// (issue #486). Mirrors the tier-resolution branches in
-/// [`resolve_tier`]: hard → no scaling; soft with a `[thresholds.soft]`
-/// table → per-metric limits (no single ratio); soft without a table →
-/// scaled by the spec's ratio (defaulting to [`DEFAULT_SOFT_HEADROOM`]).
-fn resolve_provenance(tier: TierSpec, soft_table_present: bool) -> baseline::Provenance {
-    match tier {
-        TierSpec::Hard => baseline::Provenance::hard(),
-        TierSpec::Soft(_) if soft_table_present => baseline::Provenance::soft_table(),
-        TierSpec::Soft(r) => {
-            baseline::Provenance::soft_headroom(r.unwrap_or(DEFAULT_SOFT_HEADROOM))
-        }
-    }
-}
-
-/// Resolve the effective `--format` for `check` and eagerly validate the
-/// `--output` path (#600).
-///
-/// When `--output` is given without `--format`, the format is inferred
-/// from the file extension ([`AggregatedFormat::infer_from_extension`]):
-/// `.sarif` → SARIF, `.xml` → Checkstyle. An extension with no unique
-/// writer (e.g. `.json`, shared by SARIF and Code Climate) or no
-/// extension at all is a usage error — the previous silent no-op (an
-/// explicit `--output` that wrote nothing on exit 0) was the worst CLI
-/// failure mode. When a format is in effect, an `--output` that names an
-/// existing directory is rejected before the walk rather than failing
-/// mid-write.
-///
-/// Invoking `bca check` with neither flag is left alone: the plain
-/// human stderr stream is the default contract and stays frictionless.
-///
-/// The `--output` path is only checked for being a directory; a missing
-/// parent is *not* rejected here, because the aggregated writer creates
-/// parent directories on demand (`write_to_path_or_stdout`), and that
-/// on-demand creation is a long-standing part of `check`'s contract.
-fn resolve_check_output_format(args: &mut CheckArgs) {
-    let Some(ref out) = args.output else {
-        // No `--output`: nothing to validate or infer. A bare `--format`
-        // still writes its document to stdout via the existing path.
-        return;
-    };
-
-    // The effective format: the explicit `--format`, or one inferred from
-    // the extension. An un-inferable extension is a usage error.
-    let fmt = args.output_format.unwrap_or_else(|| {
-        check_format::AggregatedFormat::infer_from_extension(out).unwrap_or_else(|| {
-            die(format_args!(
-                "--output {} has no format-bearing extension; pass --format \
-                 (checkstyle|clang-warning|code-climate|msvc-warning|sarif) \
-                 or use a recognized extension (.sarif, .xml)",
-                out.display()
-            ))
-        })
-    });
-
-    // A format is now in effect (explicit or inferred), so the path is a
-    // real sink: reject a directory up front rather than failing mid-write.
-    if out.exists() && out.is_dir() {
-        die(format_args!(
-            "--output must be a file path for `check --format {}`",
-            fmt.name()
-        ));
-    }
-
-    args.output_format = Some(fmt);
-}
-
-/// Validate `--output` / `--output-format` pairing, then resolve the
-/// effective threshold set per the documented resolution order
-/// (#373/#374/#375/#380): the manifest `[thresholds]` base, the
-/// `--config` file merged on top (keys win on collision), the tier
-/// resolution (hard verbatim, or soft via `[thresholds.soft]` /
-/// `--headroom`), and finally the absolute `--threshold` CLI overrides.
-/// Dies if no thresholds were configured. Also returns the un-scaled
-/// hard-tier limits per metric (#385) so the caller can tell a soft-band
-/// encroachment apart from a true hard breach. The set is wrapped in
-/// `Arc` so it can be cloned into each walker worker's `Config`.
-fn validate_and_build_thresholds(
-    args: &mut CheckArgs,
-    base_thresholds: ParsedThresholds,
-    tier: TierSpec,
-) -> ResolvedThresholds {
-    // Resolve the `--output` / `--format` pairing before the walk so a
-    // misconfigured invocation fails fast instead of after a full parse.
-    resolve_check_output_format(args);
-
-    // Layer 1: the manifest `[thresholds]` table (empty when no
-    // `bca.toml` was discovered). Layer 2: `--config` merges on top,
-    // its keys winning on collision, preserving every existing recipe.
-    // Both the hard and soft layers merge the same way.
-    let ParsedThresholds { mut hard, mut soft } = base_thresholds;
-    if let Some(config) = args.config.as_deref() {
-        let cfg = load_threshold_config(config);
-        hard.extend(cfg.hard);
-        soft.extend(cfg.soft);
-    }
-
-    // The soft ratio (the `RATIO` in `--tier=soft=RATIO`) was already
-    // validated to `(0, 1]` by `TierSpec::from_str` at parse time, so a
-    // typo is a clap usage error before we ever reach here.
-
-    // Capture whether a soft table is configured before `resolve_tier`
-    // borrows `soft`, so provenance resolution (#486) matches the same
-    // branch the tier resolver takes.
-    let soft_table_present = !soft.is_empty();
-
-    // Layer 3: tier resolution. Produces the per-metric limits the gate
-    // compares against. Clone `hard` so the un-scaled hard-tier limits
-    // survive for #385 hard-breach detection below.
-    let mut merged = resolve_tier(tier, hard.clone(), &soft);
-
-    // Layer 4: `--threshold` CLI flags override the resolved limit for
-    // the same metric name. They are absolute — applied *after* any
-    // scaling — because a user who typed an exact limit means it, not a
-    // fraction of it. The same value also defines the hard-tier ceiling
-    // for that metric (#385): an explicit `--threshold` is the user's
-    // declared limit, replacing whatever the hard table held.
-    for (name, limit) in &args.thresholds {
-        merged.insert(name.clone(), *limit);
-        hard.insert(name.clone(), *limit);
-    }
-    let set = ThresholdSet::build(&merged).unwrap_or_else(|e| die(e));
-    if set.is_empty() {
-        die(
-            "no thresholds configured; pass --threshold, --config, or a bca.toml [thresholds] table",
-        );
-    }
-    ResolvedThresholds {
-        set: Arc::new(set),
-        hard_limits: hard,
-        provenance: resolve_provenance(tier, soft_table_present),
-    }
-}
-
-/// Resolve the per-metric limits for the requested tier (#375/#688).
-///
-/// The soft ratio rides on the [`TierSpec`] itself — there is no longer
-/// a separate `--headroom` knob to reconcile — so the precedence model
-/// collapses to:
-///
-/// - `Hard`: the `[thresholds]` table verbatim. `[thresholds.soft]` is
-///   ignored entirely.
-/// - `Soft` with a `[thresholds.soft]` table: merge the per-metric soft
-///   overrides on top of the hard limits (metrics absent from the soft
-///   table keep their hard limit — no soft band). The blanket ratio, if
-///   one was pinned with `soft=RATIO`, is not applied: explicit
-///   per-metric limits encode intent more precisely than a multiplier.
-/// - `Soft` without a soft table: scale every hard limit by the spec's
-///   ratio (defaulting to [`DEFAULT_SOFT_HEADROOM`] for a bare `soft`).
-fn resolve_tier(
-    tier: TierSpec,
-    hard: BTreeMap<String, f64>,
-    soft: &BTreeMap<String, SoftLimit>,
-) -> BTreeMap<String, f64> {
-    let ratio = match tier {
-        TierSpec::Hard => return hard,
-        TierSpec::Soft(_) if !soft.is_empty() => {
-            // Start from the hard limits so metrics without a soft
-            // override inherit their hard limit (no soft band), then
-            // apply each soft override on top.
-            let mut out = hard;
-            for (name, soft_limit) in soft {
-                let resolved = soft_limit
-                    .resolve(name, out.get(name).copied())
-                    .unwrap_or_else(|e| die(e));
-                out.insert(name.clone(), resolved);
-            }
-            return out;
-        }
-        // No soft table: scale the hard limits by the spec's ratio,
-        // defaulting so a bare `--tier=soft` is never a silent no-op.
-        TierSpec::Soft(r) => r.unwrap_or(DEFAULT_SOFT_HEADROOM),
-    };
-    if hard.is_empty() {
-        note(
-            "--tier=soft has no effect without configured thresholds \
-             (bca.toml [thresholds] or --config); --threshold limits are \
-             absolute and are not scaled",
-        );
-    }
-    let mut out = hard;
-    for limit in out.values_mut() {
-        *limit = scale_threshold(*limit, ratio);
-    }
-    out
-}
-
-/// Run the parallel walker with a check-flavoured `Config`, collect
-/// every emitted `Violation`, and sort them by `(path, start_line,
-/// metric)` so CI diff tooling sees identical output across runs over
-/// the same tree. Returns the sorted vector plus the
-/// `files_dispatched` counter so the caller can detect the "no inputs
-/// matched" case.
-fn run_check_walk(
-    globals: GlobalOpts,
-    args: &CheckArgs,
-    preproc: Option<Arc<PreprocResults>>,
-    set: Arc<ThresholdSet>,
-) -> (Vec<Violation>, Arc<AtomicUsize>) {
-    let (tx, rx) = std::sync::mpsc::channel();
-    let files_dispatched = Arc::new(AtomicUsize::new(0));
-    let cfg = Config {
-        threshold_set: Some(set),
-        check_tx: Some(Mutex::new(tx)),
-        files_dispatched: Some(Arc::clone(&files_dispatched)),
-        suppression_policy: SuppressionPolicy::from_no_suppress(args.no_suppress),
-        report_suppressed: args.report_suppressed,
-        // Compute body hashes during the walk only when fuzzy matching
-        // is requested — whether for a `--baseline` read or to populate
-        // `body_hash` in a `--write-baseline` write.
-        fuzzy_baseline: args.baseline_fuzzy_match.unwrap_or(false),
-        ..Config::new(Action::Check, &globals, preproc)
-    };
-    run_walk(globals, cfg);
-
-    // Workers have all joined by the time `run_walk` returns, so the
-    // sender side is dropped and `rx.into_iter()` terminates cleanly.
-    let mut violations: Vec<Violation> = rx.into_iter().collect();
-    // Stable, deterministic stderr output: by path, then start line, then
-    // metric name. Different runs over the same tree produce identical
-    // output, which CI diff tooling relies on.
-    violations.sort_by(|a, b| {
-        a.path
-            .cmp(&b.path)
-            .then(a.start_line.cmp(&b.start_line))
-            .then(a.metric.cmp(b.metric))
-    });
-
-    (violations, files_dispatched)
-}
-
-/// Serialize and write the collected violations as a baseline TOML
-/// file. Used by the `--write-baseline` early-exit branch. The
-/// baseline-file directory becomes the *anchor* — every entry's path
-/// is keyed relative to it, so a subsequent `--baseline` invocation
-/// from any `--paths` form (`.`, `src/`, `$PWD`) still matches.
-fn write_check_baseline(violations: Vec<Violation>, path: &Path, provenance: baseline::Provenance) {
-    let anchor = baseline::anchor_for(path);
-    let file = baseline::from_violations(violations, &anchor, provenance);
-    let entry_count = file.entries.len();
-    let text =
-        baseline::render(&file).unwrap_or_else(|e| die(format_args!("serialize baseline: {e}")));
-    write_atomic(path, text.as_bytes()).unwrap_or_else(|e| die_io("write baseline", path, e));
-    eprintln!(
-        "bca: wrote {entry_count} baseline entries to {}",
-        path.display()
-    );
-}
-
-fn apply_check_exclude(
-    violations: Vec<Violation>,
-    args: &CheckArgs,
-    paths: &[PathBuf],
-    paths_from: Option<&Path>,
-) -> Vec<Violation> {
-    // Fast path: nothing configured (the common case) skips the
-    // glob-set build, the `--paths-from` re-read, and the file read
-    // entirely.
-    if args.check_exclude.is_empty() && args.check_exclude_from.is_none() {
-        return violations;
-    }
-    let globset = crate::build_exclude_globset(
-        args.check_exclude.clone(),
-        args.check_exclude_from.as_deref(),
-        "--check-exclude-from",
-    );
-    // Anchor each violation's emitted path to the walk-root `./`-form
-    // before matching, mirroring the global `--exclude`/`--include`
-    // anchoring (#489), so a `./`-anchored `[check.exclude]` pattern
-    // exempts the same files regardless of how the seed resolved
-    // (absolute, `$PWD`, or a manifest root above the CWD).
-    //
-    // The seeds must be the *full* set the walk anchored against —
-    // `--paths` AND any `--paths-from` entries — reanchored exactly as
-    // [`expand_seed_paths`] did. Threading only `--paths` left a
-    // violation from a `--paths-from`-sourced (e.g. absolute) seed
-    // matched unanchored, so a `[check.exclude]` glob silently failed to
-    // exempt it (#497). Re-reading `--paths-from` here (rather than
-    // plumbing the resolved list back from the walk) keeps the hot
-    // no-exclude path above allocation-free; this branch is the rare
-    // configured case.
-    let mut seeds: Vec<PathBuf> = paths.to_vec();
-    if let Some(src) = paths_from {
-        seeds.extend(crate::read_paths_from(src).unwrap_or_else(|e| die(e)));
-    }
-    let seeds: Vec<PathBuf> = seeds
-        .into_iter()
-        .map(crate::walk_seed::reanchor_seed)
-        .collect();
-    let before = violations.len();
-    let kept: Vec<Violation> = violations
-        .into_iter()
-        .filter(|v| {
-            // Anchor to the `./`-relative walk-root form, then strip the
-            // leading `./` so bare-relative `[check.exclude]` patterns match
-            // it just like `./`-prefixed ones (#726).
-            let anchored = crate::walk_seed::anchor_against_seeds(&seeds, &v.path);
-            !globset.is_match(crate::walk_seed::strip_cur_dir(&anchored))
-        })
-        .collect();
-    let skipped = before - kept.len();
-    if skipped > 0 {
-        eprintln!("bca: skipped {skipped} violations via [check.exclude]");
-    }
-    kept
-}
-
-/// Classify each violation against the optional `--baseline` file.
-/// The kept list carries `(Violation, Option<Coverage>)` so the
-/// stderr renderer can attach a `[new]` / `[regr +N%]` tag. Without
-/// `--baseline`, `Option<Coverage>` is `None` and the renderer emits
-/// the exact pre-tag line format byte-identically.
-/// Compose the stderr warning (issue #486) when the current run is
-/// stricter than the baseline was written against, or `None` when the
-/// comparison is safe (see [`baseline::check_provenance`] for the
-/// directional rule). Split out from [`filter_by_baseline`] so a test
-/// can pin the exact message and the silent cases without a baseline
-/// file on disk.
-fn provenance_warning(
-    current: baseline::Provenance,
-    baseline: Option<baseline::Provenance>,
-) -> Option<String> {
-    match baseline::check_provenance(current, baseline) {
-        baseline::ProvenanceCheck::Ok => None,
-        baseline::ProvenanceCheck::StricterThanBaseline {
-            current: cur,
-            baseline: base,
-        } => Some(format!(
-            "this check's effective limits (strictness {cur}) are \
-             stricter than the baseline was written against (strictness \
-             {base}); the baseline may under-cover and the gate can fire on \
-             untouched files. Refresh it at the matching tier, e.g. \
-             `bca check --tier=soft={cur} --write-baseline \
-             <file>` (or `--write-baseline <file>` for the hard tier)."
-        )),
-    }
-}
-
-fn filter_by_baseline(
-    violations: Vec<Violation>,
-    baseline_path: Option<&Path>,
-    tolerance: usize,
-    fuzzy: bool,
-    provenance: baseline::Provenance,
-    keep_covered: bool,
-) -> Vec<(Violation, Option<Coverage>)> {
-    let Some(path) = baseline_path else {
-        return violations.into_iter().map(|v| (v, None)).collect();
-    };
-    let baseline = load_baseline(path, tolerance, fuzzy);
-    // Issue #486: warn when this run's effective limits are stricter than
-    // the baseline was written against (the baseline may under-cover and
-    // the gate can fire on untouched files). Silent in the safe
-    // directions (hard reading soft, equal, absent provenance).
-    if let Some(msg) = provenance_warning(provenance, baseline.provenance()) {
-        warn(msg);
-    }
-    let before = violations.len();
-    let kept: Vec<_> = violations
-        .into_iter()
-        .filter_map(|v| match baseline.classify(&v) {
-            // `--report-suppressed` keeps baseline-covered offenders (tagged
-            // `Covered`) so they can be surfaced as `external` suppressions
-            // in the document; the split in `run_check` keeps them out of the
-            // gate. The default path still drops them entirely.
-            Coverage::Covered { .. } if !keep_covered => None,
-            c => Some((v, Some(c))),
-        })
-        .collect();
-    let filtered = before - kept.len();
-    if filtered > 0 {
-        eprintln!("bca: filtered {filtered} violations via baseline");
-    }
-    kept
-}
-
-/// Resolve the diff scope for `--since` / `--changed-only` /
-/// auto-detected env vars. Behaviour:
-///
-/// - No flag, no env signal → `None`. The footer prints today's
-///   single-section listing; `--changed-only` is rejected at the
-///   top of the helper because it requires a scope.
-/// - Resolved scope (`ResolveOutcome::Ok`) → `Some(scope)`, surfaced
-///   in the footer banner and used to bucket touched-in-range rows.
-/// - Resolver hit an error (`ResolveOutcome::Failed`) → fatal when
-///   `--changed-only` is passed (otherwise the gate would silently
-///   suppress nothing), warning-only otherwise (the developer still
-///   sees the offender list, just without the touched-in-range
-///   partition).
-fn resolve_diff_scope(args: &CheckArgs) -> Option<diff::DiffScope> {
-    let outcome = diff::resolve_scope(args.since.as_deref());
-    match outcome {
-        diff::ResolveOutcome::Ok(scope) => Some(scope),
-        diff::ResolveOutcome::Disabled => {
-            if args.changed_only {
-                die("--changed-only requires --since <ref> or one of \
-                     BCA_DIFF_BASE / GITHUB_BASE_REF / GITHUB_EVENT_BEFORE \
-                     in the environment");
-            }
-            None
-        }
-        diff::ResolveOutcome::Failed { reason, source } => {
-            if args.changed_only {
-                die(format_args!(
-                    "--changed-only: failed to resolve diff base via {}: {reason}",
-                    source.label(),
-                ));
-            }
-            eprintln!(
-                "bca: --since/auto-detect via {} failed ({reason}); proceeding without diff scope",
-                source.label(),
-            );
-            None
-        }
-    }
-}
-
-fn apply_changed_only(
-    pairs: Vec<(Violation, Option<Coverage>)>,
-    scope: Option<&diff::DiffScope>,
-    changed_only: bool,
-) -> Vec<(Violation, Option<Coverage>)> {
-    let outcome = apply_changed_only_inner(pairs, scope, changed_only);
-    if let Some(diag) = outcome.diagnostic {
-        eprintln!("{diag}");
-    }
-    outcome.kept
-}
-
-/// Result of [`apply_changed_only_inner`]: the filtered pairs plus
-/// an optional diagnostic string for the caller to surface. Extracted
-/// from the outer `apply_changed_only` so tests can pin the
-/// diagnostic shape (the "silent regression" guard the audit-tests
-/// pass would otherwise miss).
-struct ChangedOnlyOutcome {
-    kept: Vec<(Violation, Option<Coverage>)>,
-    diagnostic: Option<String>,
-}
-
-fn apply_changed_only_inner(
-    pairs: Vec<(Violation, Option<Coverage>)>,
-    scope: Option<&diff::DiffScope>,
-    changed_only: bool,
-) -> ChangedOnlyOutcome {
-    if !changed_only {
-        return ChangedOnlyOutcome {
-            kept: pairs,
-            diagnostic: None,
-        };
-    }
-    let Some(scope) = scope else {
-        // `resolve_diff_scope` fatal-errors when `--changed-only` is
-        // set without a resolvable scope, so this branch is
-        // unreachable from the production `run_check` pipeline. It
-        // exists for tests and to defend against a future refactor
-        // that bypasses that gate — degrade to a no-op rather than
-        // silently emit the empty set (which would green-light the
-        // gate on a misconfigured CI), but log so the operator
-        // notices.
-        return ChangedOnlyOutcome {
-            kept: pairs,
-            diagnostic: Some(
-                "bca: --changed-only requested but no diff scope is available; \
-                 skipping filter (would-be programmer error — \
-                 resolve_diff_scope should have fatal-errored)"
-                    .to_string(),
-            ),
-        };
-    };
-    if scope.changed.is_empty() {
-        // A resolved-but-empty scope (e.g. running `--since main` from
-        // a branch that has been merged/squashed into main locally, or
-        // a force-pushed branch where the diff base now equals HEAD)
-        // would otherwise silently drop every violation and exit 0,
-        // which is exactly the "silent green-light" failure mode #359
-        // was meant to prevent. Surface it explicitly so CI logs make
-        // it obvious the gate ran but had nothing to check. Branch on
-        // `pairs.is_empty()` so the wording matches reality: "dropping
-        // 0 violations" would suggest the gate suppressed something
-        // it did not, confusing a developer reading a clean PR log.
-        let diag = if pairs.is_empty() {
-            format!(
-                "bca: --changed-only: diff scope is empty (no files touched between {} and HEAD); \
-                 no violations to check and no files in diff scope",
-                scope.base,
-            )
-        } else {
-            format!(
-                "bca: --changed-only: diff scope is empty (no files touched between {} and HEAD); \
-                 dropping {} violations and exiting clean",
-                scope.base,
-                pairs.len()
-            )
-        };
-        return ChangedOnlyOutcome {
-            kept: Vec::new(),
-            diagnostic: Some(diag),
-        };
-    }
-    // Memoize `scope.contains` (which canonicalizes internally) by
-    // raw `v.path`. Real-world inputs cluster heavily per file
-    // (a 50-violation run typically touches 5-10 files), so this
-    // turns O(violations) realpath(2) calls into O(unique raw
-    // paths). Precondition: the walker must emit violation paths in
-    // a canonical-form-consistent style across one check run (it
-    // does — paths are always rooted at the same `--paths` seed
-    // and don't mix `./X` with `X`). If a future change introduces
-    // alias paths in a single run, two violations of the same file
-    // would each pay a separate canonicalize call — the cache would
-    // still be correct, just not optimal.
-    let mut in_scope: HashMap<PathBuf, bool> = HashMap::new();
-    let before = pairs.len();
-    let kept: Vec<_> = pairs
-        .into_iter()
-        .filter(|(v, _)| {
-            *in_scope
-                .entry(v.path.clone())
-                .or_insert_with(|| scope.contains(&v.path))
-        })
-        .collect();
-    let dropped = before - kept.len();
-    let diagnostic = (dropped > 0).then(|| {
-        format!("bca: --changed-only dropped {dropped} violations from files outside diff scope")
-    });
-    ChangedOnlyOutcome { kept, diagnostic }
-}
-
-fn emit_check_results(
-    pairs: Vec<(Violation, Option<Coverage>)>,
-    suppressed: Vec<(Violation, Option<Coverage>)>,
-    args: &CheckArgs,
-    scope: Option<&diff::DiffScope>,
-    remediation: Option<&str>,
-) {
-    // BrokenPipe on stderr (e.g. when piped to `head`) is the only
-    // realistic write failure here; swallow it rather than die so the
-    // exit-code contract is honored.
-    let mut stderr = std::io::stderr().lock();
-    for (v, tag) in &pairs {
-        let _ = writeln!(stderr, "{}", render_violation_line(v, tag.as_ref()));
-    }
-    if !args.no_summary && !pairs.is_empty() {
-        let _ = write_summary_footer(&mut stderr, &pairs, scope);
-    }
-    if github_annotations_enabled(args) && !pairs.is_empty() {
-        // Emit annotations *after* the human stream + summary footer
-        // so a reader tailing the CI log sees the contiguous
-        // human-readable block first. The GHA log viewer scrapes
-        // `::error…` lines wherever they appear and renders them as
-        // inline annotations on the file-diff view regardless of
-        // position.
-        let _ = check_format::write_github_annotations(
-            &mut stderr,
-            pairs.iter().map(|(v, _)| v),
-            check_format::DEFAULT_GITHUB_ANNOTATION_CAP,
-        );
-    }
-    // The remediation block is the final thing on stderr — a reader
-    // skimming a CI log sees it as the natural "what now?" answer
-    // immediately after the failure evidence. Skipped when the
-    // caller passed `None` (clean run, or `--no-remediation`).
-    if let Some(block) = remediation {
-        let _ = write!(stderr, "{block}");
-    }
-    drop(stderr);
-
-    // Append the markdown digest to `$GITHUB_STEP_SUMMARY` (or the
-    // user-supplied `--summary-file`). Writes are bracketed by the
-    // bca-step-summary markers so a retried GHA step replaces
-    // (instead of stacks) the previous block. Failures here are
-    // logged but never affect the exit-code contract — the
-    // step-summary panel is informational.
-    if let Some(path) = step_summary_path(args)
-        && let Err(e) = check_format::write_step_summary(&path, &pairs, remediation)
-    {
-        eprintln!(
-            "bca: failed to append step summary to {}: {e}",
-            path.display()
-        );
-    }
-
-    // Emit the aggregated CI/IDE document if requested. Empty input
-    // produces a well-formed but offender-free document, which CI
-    // consumers can ingest unchanged on clean runs. The exit-code
-    // contract is unaffected by this branch.
-    if let Some(fmt) = args.output_format {
-        let offenders: Vec<_> = pairs
-            .into_iter()
-            .map(|(v, _)| violation_to_offender(v))
-            .collect();
-        // Only the SARIF format can represent suppression, so route active +
-        // suppressed offenders through the suppression-aware writer there. For
-        // every other format (and the default no-suppressed case) fall back to
-        // the plain dump so output is byte-for-byte unchanged.
-        let written = if !suppressed.is_empty()
-            && matches!(fmt, check_format::AggregatedFormat::Sarif)
-        {
-            check_format::dump_sarif_with_suppressed(&offenders, suppressed, args.output.as_deref())
-        } else {
-            fmt.dump(&offenders, args.output.as_deref())
-        };
-        written.unwrap_or_else(|e| die(format_args!("failed to write {}: {e}", fmt.name())));
-    }
-}
-
-/// Severity category of a `bca check` run, used to derive the process
-/// exit code (#385). The variants are *not* the exit codes — the
-/// mapping depends on `--strict-exit-codes` (see [`Self::exit_code`]).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum CheckOutcome {
-    /// No violations survived filtering.
-    Clean,
-    /// Violations exist, but none is a baseline regression and none
-    /// breaches the hard limit under `--tier=soft`. Also the bucket for
-    /// every violation when no `--baseline` was supplied (nothing is
-    /// baselined, so nothing can have "regressed").
-    NewOnly,
-    /// Every kept violation matched a baseline entry that worsened.
-    RegressionOnly,
-    /// A mix of new offenders and baseline regressions.
-    Mixed,
-    /// At least one `--tier=soft` violation also exceeds the hard
-    /// limit — escalated above the new/regression split because a true
-    /// breach is more urgent than soft-band encroachment.
-    HardBreach,
-}
-
-impl CheckOutcome {
-    /// Map the outcome to a process exit code. In the default contract
-    /// (`strict == false`) every non-clean run collapses to exit `2`,
-    /// preserving the stable 0/1/2 behaviour every existing integration
-    /// relies on. In tiered mode (`--strict-exit-codes`) each category
-    /// gets its own code (2-5). Returns `None` for a clean run, where
-    /// the caller exits 0 implicitly by returning.
-    fn exit_code(self, strict: bool) -> Option<i32> {
-        let tiered = match self {
-            Self::Clean => return None,
-            Self::NewOnly => 2,
-            Self::RegressionOnly => 3,
-            Self::Mixed => 4,
-            Self::HardBreach => 5,
-        };
-        // The default contract collapses every violation category to
-        // exit 2; only `--strict-exit-codes` surfaces the 3/4/5 split.
-        Some(if strict { tiered } else { 2 })
-    }
-}
-
-/// Categorise the kept violations for the exit-code contract (#385).
-///
-/// `hard_limits` holds the resolved hard-tier limit per metric. It is
-/// consulted only at the soft tier, where a violation whose value also
-/// exceeds the hard limit escalates to [`CheckOutcome::HardBreach`]. At
-/// the hard tier every violation already exceeds the hard limit, so the
-/// escalation is suppressed (it would otherwise swallow the new/regr
-/// split) and only baseline coverage drives the result.
-fn classify_check_outcome(
-    pairs: &[(Violation, Option<Coverage>)],
-    tier: Tier,
-    hard_limits: &BTreeMap<String, f64>,
-) -> CheckOutcome {
-    if pairs.is_empty() {
-        return CheckOutcome::Clean;
-    }
-    let mut has_new = false;
-    let mut has_regression = false;
-    let mut has_hard_breach = false;
-    for (v, coverage) in pairs {
-        // Direction-aware hard-breach escalation (#837): mirror the
-        // gate's breach test so the lower-is-worse `mi.*` family escalates
-        // on a value *below* the hard floor, not above it. A NaN value
-        // (degenerate Halstead on a trivial function) fails both
-        // comparisons in `breaches_limit`, so it never escalates to a hard
-        // breach; it falls to the new/regr split below, mirroring how
-        // `Baseline::classify` treats a NaN as `Regressed` rather than a
-        // magnitude. A NaN has no meaningful distance from the ceiling.
-        if tier == Tier::Soft
-            && let Some(&hard) = hard_limits.get(v.metric)
-            && breaches_limit(v.value, hard, v.lower_is_worse)
-        {
-            has_hard_breach = true;
-        }
-        match coverage {
-            Some(Coverage::Regressed { .. }) => has_regression = true,
-            // `Coverage::New`, or `None` when no `--baseline` was given.
-            // `Coverage::Covered` never reaches here: by default
-            // `filter_by_baseline` drops it, and under `--report-suppressed`
-            // `run_check` partitions it into the `suppressed` set — so the
-            // `active` slice this fn classifies is always Covered-free.
-            _ => has_new = true,
-        }
-    }
-    if has_hard_breach {
-        CheckOutcome::HardBreach
-    } else if has_new && has_regression {
-        CheckOutcome::Mixed
-    } else if has_regression {
-        CheckOutcome::RegressionOnly
-    } else {
-        CheckOutcome::NewOnly
-    }
-}
-
-/// Decide whether GitHub Actions `::error` annotations should be
-/// emitted (issue #683). The tri-state `--github-annotations
-/// <auto|always|never>` flag resolves like `--color`: `always` forces
-/// them on, `never` suppresses them even inside a workflow step, and
-/// `auto` (the default) falls back to `$GITHUB_ACTIONS == "true"`, the
-/// signal GHA sets inside every step.
-fn github_annotations_enabled(args: &CheckArgs) -> bool {
-    let in_gha = std::env::var(check_format::GITHUB_ACTIONS_ENV).as_deref() == Ok("true");
-    args.github_annotations.enabled_with(in_gha)
-}
-
-/// Resolve the path to append the step-summary digest to (issue #683).
-/// `--summary-file <path>` appends there unconditionally; `--summary-file
-/// never` suppresses the digest even inside a GHA step; `auto` (the
-/// default when the flag is omitted) defers to `$GITHUB_STEP_SUMMARY`.
-/// Returns `None` when no path is in effect and the digest is skipped.
-fn step_summary_path(args: &CheckArgs) -> Option<PathBuf> {
-    match &args.summary_file {
-        Some(SummaryFile::Path(p)) => Some(p.clone()),
-        Some(SummaryFile::Never) => None,
-        // `auto` (explicit) and the unset default both detect the env var.
-        Some(SummaryFile::Auto) | None => {
-            std::env::var_os(check_format::GITHUB_STEP_SUMMARY_ENV).map(PathBuf::from)
-        }
-    }
-}
-
-fn format_remediation_block(globals: &GlobalOpts, args: &CheckArgs) -> Option<String> {
-    use std::fmt::Write as _;
-    if args.no_remediation {
-        return None;
-    }
-    let mut out = String::from("\n--- next steps ---\n");
-    let _ = writeln!(out, "* Detailed reports: {}", artifact_link());
-    let _ = writeln!(
-        out,
-        "* To refresh baseline: {}",
-        refresh_baseline_command(globals, args)
-    );
-    // The refresh command mirrors path filters (`--paths`,
-    // `--exclude`, `--exclude-from`, `--config`, `--baseline`) but
-    // intentionally omits selectors that don't affect baseline
-    // composition (`--num-jobs`) and ones that would bloat the
-    // common-case command (`--include`, `--language`,
-    // `--paths-from`, `--exclude-tests`). Surface the omission so a
-    // user with a non-trivial scope re-adds them rather than
-    // assuming the printed command is complete.
-    out.push_str(
-        "  (mirrors path filters only — re-add any --include / --language / --exclude-tests / --paths-from flags as needed)\n",
-    );
-    out.push_str(
-        "* Adoption guide: https://dekobon.github.io/big-code-analysis/recipes/baselines.html\n",
-    );
-    Some(out)
-}
-
-fn artifact_link() -> String {
-    artifact_link_for(
-        std::env::var(check_format::GITHUB_REPOSITORY_ENV).ok(),
-        std::env::var(check_format::GITHUB_RUN_ID_ENV).ok(),
-    )
-}
-
-/// Pure inner: render the detailed-reports bullet given explicit env
-/// values (rather than reading them from the process environment).
-/// Extracted so tests can pin both the SOME and NONE branches without
-/// depending on whether the test process happens to have GHA env
-/// vars set. Empty strings are treated as absent — GitHub Actions
-/// does set these vars but the spec doesn't promise non-empty values
-/// on every event type.
-///
-/// Only when both env vars are present (i.e. running inside a GitHub
-/// Actions job) do we point at the uploaded `bca-reports` artifact;
-/// outside CI there is no run and no artifact, so claiming one sends
-/// developers hunting for an upload that does not exist (#676). The
-/// local fallback suggests `bca report` for the detailed view.
-fn artifact_link_for(repo: Option<String>, run_id: Option<String>) -> String {
-    let repo = repo.filter(|s| !s.is_empty());
-    let run_id = run_id.filter(|s| !s.is_empty());
-    match (repo, run_id) {
-        (Some(repo), Some(run_id)) => {
-            format!("bca-reports artifact at https://github.com/{repo}/actions/runs/{run_id}")
-        }
-        _ => "run `bca report` to see them locally".to_string(),
-    }
-}
-
-fn refresh_baseline_command(globals: &GlobalOpts, args: &CheckArgs) -> String {
-    let mut cmd = String::from("bca");
-    let paths: Vec<&Path> = if globals.paths.is_empty() {
-        // Default-when-absent — mirror the walker's `expand_seed_paths`
-        // `.` fallback (#596) so the printed refresh command behaves
-        // identically to the pathless invocation that produced it.
-        vec![Path::new(".")]
-    } else {
-        globals.paths.iter().map(PathBuf::as_path).collect()
-    };
-    for p in &paths {
-        cmd.push_str(" --paths ");
-        cmd.push_str(&shell_quote_path(p));
-    }
-    for ex in &globals.exclude {
-        cmd.push_str(" --exclude ");
-        cmd.push_str(&shell_quote(ex));
-    }
-    if let Some(p) = &globals.exclude_from {
-        cmd.push_str(" --exclude-from ");
-        cmd.push_str(&shell_quote_path(p));
-    }
-    cmd.push_str(" check");
-    if let Some(p) = &args.config {
-        cmd.push_str(" --config ");
-        cmd.push_str(&shell_quote_path(p));
-    }
-    // `--baseline` and `--write-baseline` conflict in clap, so we
-    // prefer the user's baseline path if they ran with it (the
-    // refresh writes back to the same file). Fall back to the
-    // documented default `.bca-baseline.toml`.
-    cmd.push_str(" --write-baseline ");
-    match args.baseline.as_deref() {
-        Some(p) => cmd.push_str(&shell_quote_path(p)),
-        None => cmd.push_str(&shell_quote(".bca-baseline.toml")),
-    }
-    cmd
-}
-
-fn shell_quote_path(p: &Path) -> String {
-    // The printed command is an *identifier* in the user's shell —
-    // running it must reach the same file `bca` walked. Non-UTF-8
-    // paths cannot be expressed as a shell argument verbatim, so
-    // surface them as a clearly-broken placeholder rather than emit
-    // a `to_string_lossy` form that silently points at the wrong
-    // file (AGENTS.md: identifier paths use `to_str()` with explicit
-    // non-UTF-8 handling, not `path.display()`).
-    //
-    // The placeholder contains `<`, `>` and spaces, which force
-    // `shell_quote`'s slow path → single-quoted literal. Combining
-    // the to_str + quote here (instead of leaving callers to chain
-    // them) makes the discipline structural: a future caller can't
-    // accidentally `eprintln!("{}", path_for_shell(p))` and emit an
-    // unquoted `<non-UTF-8 path: …>` that bash would parse as input
-    // redirection.
-    let raw = p.to_str().map_or_else(
-        || format!("<non-UTF-8 path: {}>", p.display()),
-        str::to_string,
-    );
-    shell_quote(&raw)
-}
-
-/// Shell-quote `s` for inclusion in the remediation block's
-/// copy-paste command. Uses single-quoting for simplicity: every
-/// character is literal inside `'...'` except `'` itself, which we
-/// escape via `'\''`. ASCII-safe and POSIX-compatible.
-///
-/// **POSIX-only**: This quoting is correct for bash / zsh / dash /
-/// sh, which is what GitHub Actions runs every step in. It is NOT
-/// safe for `cmd.exe` or `PowerShell` — a Windows user copy-pasting
-/// the refresh command from a Windows CI log would need to
-/// re-escape. The remediation block is a GHA/POSIX-CI feature by
-/// design; Windows-host CI is out of scope.
-fn shell_quote(s: &str) -> String {
-    // Fast path: identifiers / paths without metacharacters need no
-    // quoting at all. Keeping them unquoted makes the copy-paste
-    // command read naturally for the common case.
-    if !s.is_empty()
-        && s.chars().all(|c| {
-            c.is_ascii_alphanumeric() || matches!(c, '_' | '-' | '.' | '/' | ':' | '=' | ',' | '@')
-        })
-    {
-        return s.to_string();
-    }
-    let mut out = String::with_capacity(s.len() + 2);
-    out.push('\'');
-    for ch in s.chars() {
-        if ch == '\'' {
-            out.push_str("'\\''");
-        } else {
-            out.push(ch);
-        }
-    }
-    out.push('\'');
-    out
-}
-
-struct FooterRow<'a> {
-    count: usize,
-    worst: &'a Violation,
-    display: String,
-    path: &'a Path,
-}
-
-fn compute_footer_rows(pairs: &[(Violation, Option<Coverage>)]) -> Vec<FooterRow<'_>> {
-    Violation::group_pairs_by_path(pairs)
-        .into_iter()
-        .map(|(count, worst, display, path)| FooterRow {
-            count,
-            worst,
-            display,
-            path,
-        })
-        .collect()
-}
-
-/// Emit each row in `rows`, propagating the first I/O error. Used
-/// by both the legacy single-section path and the per-bucket
-/// partitioned path so the row format stays in lockstep.
-fn emit_footer_rows(w: &mut impl Write, rows: &[FooterRow<'_>]) -> std::io::Result<()> {
-    for row in rows {
-        write_footer_row(w, row.count, row.worst, &row.display)?;
-    }
-    Ok(())
-}
-
-/// Emit the "Files in this range:" header followed by the touched
-/// rows. When the diff scope had no offenders in it, emit an
-/// explicit "(none — …)" line so the reader gets a positive "your
-/// change is clean" signal instead of having to compare both halves
-/// of the footer to confirm absence.
-fn write_in_range_section(
-    w: &mut impl Write,
-    scope: &diff::DiffScope,
-    in_range: &[FooterRow<'_>],
-) -> std::io::Result<()> {
-    writeln!(
-        w,
-        "Files in this range (diff base: {} via {}):",
-        scope.base,
-        scope.source.label()
-    )?;
-    if in_range.is_empty() {
-        writeln!(w, "  (none — no offenders in files touched by this diff)")?;
-    } else {
-        emit_footer_rows(w, in_range)?;
-    }
-    Ok(())
-}
-
-/// Emit the "Other offenders:" header followed by the legacy
-/// offender list (files not touched by the diff scope). Returns a
-/// clean `Ok(())` when `other` is empty so the caller need not gate
-/// the call — the section's heading would be misleading without
-/// rows below it.
-fn write_other_section(w: &mut impl Write, other: &[FooterRow<'_>]) -> std::io::Result<()> {
-    if other.is_empty() {
-        return Ok(());
-    }
-    writeln!(w)?;
-    writeln!(w, "Other offenders:")?;
-    emit_footer_rows(w, other)
-}
-
-fn write_summary_footer(
-    w: &mut impl Write,
-    pairs: &[(Violation, Option<Coverage>)],
-    scope: Option<&diff::DiffScope>,
-) -> std::io::Result<()> {
-    // The caller (`emit_check_results`) gates on `!pairs.is_empty()`,
-    // so `compute_footer_rows` should always return at least one
-    // row. Assert in debug builds so a future refactor that
-    // surfaces the footer on clean runs (e.g. for positive-
-    // confirmation symmetry with the step-summary "✓ No threshold
-    // violations" message) doesn't silently emit a dangling
-    // `Files in this range:` banner with no body.
-    let rows = compute_footer_rows(pairs);
-    debug_assert!(
-        !rows.is_empty(),
-        "write_summary_footer called with no rows; \
-         caller must gate on !pairs.is_empty()"
-    );
-    writeln!(w)?;
-    writeln!(w, "--- summary ---")?;
-    let Some(s) = scope else {
-        // Without a scope, today's single-section footer is
-        // byte-identical to the pre-#359 output. This is the
-        // load-bearing back-compat path for CI tooling that grep-
-        // anchors on the legacy footer shape.
-        return emit_footer_rows(w, &rows);
-    };
-    // With a scope, partition rows into "touched in this range" vs
-    // legacy offenders. `DiffScope::contains` canonicalises once per
-    // row group (already deduplicated by `compute_footer_rows`), so
-    // the partitioning is at worst O(unique files) realpath(2) calls.
-    let (in_range, other): (Vec<_>, Vec<_>) =
-        rows.into_iter().partition(|row| s.contains(row.path));
-    write_in_range_section(w, s, &in_range)?;
-    write_other_section(w, &other)
-}
-
-/// Render a single per-file footer row. Shared between the in-range
-/// and other-offenders sections so the formatting stays in lockstep.
-fn write_footer_row(
-    w: &mut impl Write,
-    count: usize,
-    worst: &Violation,
-    display: &str,
-) -> std::io::Result<()> {
-    let noun = if count == 1 {
-        "violation"
-    } else {
-        "violations"
-    };
-    writeln!(
-        w,
-        "{display}: {count} {noun} (worst: {} = {} vs limit {} at L{})",
-        worst.metric,
-        MetricScalar(worst.value),
-        MetricScalar(worst.limit),
-        worst.start_line,
-    )
-}
+pub(crate) use {
+    analyze::*, check::*, diff_cmd::*, exemptions::*, init::*, preproc::*, report::*, scan::*,
+};
 
 /// Parse `std::env::args_os()` and execute the selected `bca`
 /// subcommand. Intended to be called from the `bca` binary's `main`,
@@ -1657,1260 +298,6 @@ fn num_jobs_set_on_cli(matches: &ArgMatches) -> bool {
     match matches.subcommand() {
         Some((_, sub)) => num_jobs_set_on_cli(sub),
         None => false,
-    }
-}
-
-fn run_command_list_metrics(args: ListMetricsArgs) {
-    let mut buf = Vec::new();
-    write_metrics(&mut buf, args.mode).expect("writing to Vec<u8> is infallible");
-    write_stdout_or_die(&buf);
-}
-
-fn run_command_dump(globals: GlobalOpts, line: LineRange, preproc: Option<Arc<PreprocResults>>) {
-    // `dump` is the documented exception to #596's default-`.` walk: a
-    // whole-tree AST dump of the cwd has no plausible use and scrolls
-    // thousands of interleaved lines, so bare `bca dump` errors asking
-    // for a path instead of defaulting to `.` (#690). A path may arrive
-    // via `--paths`, `--paths-from`, or a manifest `paths` key (all
-    // merged into `globals.paths` / `paths_from` before we reach here).
-    if globals.paths.is_empty() && globals.paths_from.is_none() {
-        die("dump needs an explicit path: pass one with --paths <PATH> \
-             (a whole-tree AST dump of the current directory is never useful)");
-    }
-    let cfg = Config {
-        line_start: line.line_start,
-        line_end: line.line_end,
-        ..Config::new(Action::Dump, &globals, preproc)
-    };
-    run_walk(globals, cfg);
-}
-
-fn run_command_functions(globals: GlobalOpts, preproc: Option<Arc<PreprocResults>>) {
-    let cfg = Config::new(Action::Functions, &globals, preproc);
-    run_walk(globals, cfg);
-}
-
-/// Resolved structured-output destination for `metrics` / `ops` (#669):
-/// stdout, a single aggregate file (`--output <FILE>`), or a per-file
-/// directory tree (`--output-dir <DIR>`).
-enum StructuredOutput {
-    Stdout,
-    AggregateFile(PathBuf),
-    Dir(PathBuf),
-}
-
-/// Validate the `--output` / `--output-dir` / `--format` combination for
-/// `metrics` / `ops` and resolve the destination (#669).
-///
-/// `--output <FILE>` now means a single aggregate file everywhere;
-/// `--output-dir <DIR>` carries the per-file-tree mode `--output` used to
-/// imply. The guards enforced here:
-///
-/// - `--output` and `--output-dir` together → error (one destination
-///   only).
-/// - Either destination without a structured `--format` → error (#661):
-///   the default `text` format streams a human-readable tree to stdout
-///   and writes no files, so a destination under it would silently no-op.
-/// - CBOR with no destination → error (binary cannot go to stdout).
-/// - `--output` must be a writable file path (parent exists, not an
-///   existing directory); `--output-dir` must not name an existing
-///   non-directory.
-///
-/// `command` names the subcommand for the messages.
-fn resolve_structured_output(
-    have_format: bool,
-    output: Option<PathBuf>,
-    output_dir: Option<PathBuf>,
-    command: &str,
-) -> StructuredOutput {
-    if output.is_some() && output_dir.is_some() {
-        die(format_args!(
-            "`{command}`: --output (single file) and --output-dir (per-file \
-             directory) are mutually exclusive; pass only one"
-        ));
-    }
-    // #661: a destination without a structured format silently no-ops
-    // under the default text stream. Error loudly so CI never consumes a
-    // stale/missing artifact.
-    if !have_format && (output.is_some() || output_dir.is_some()) {
-        die(format_args!(
-            "`{command} --output`/`--output-dir` needs a structured format: \
-             the default text format streams to stdout and writes no files. \
-             Pass --format json|yaml|toml|cbor|csv."
-        ));
-    }
-    if let Some(out) = output {
-        validate_output_path(&out, command);
-        return StructuredOutput::AggregateFile(out);
-    }
-    if let Some(dir) = output_dir {
-        if dir.exists() && !dir.is_dir() {
-            die(format_args!(
-                "--output-dir must be a directory for `{command}`"
-            ));
-        }
-        return StructuredOutput::Dir(dir);
-    }
-    StructuredOutput::Stdout
-}
-
-/// Resolve the `--metrics` selection (#691) into the library `Metric`
-/// families to compute, or `None` (compute all metrics) when the flag is
-/// absent. Every requested name is validated against the catalog via the
-/// shared #662 did-you-mean validator first, so an unrecognized name is a
-/// hard error (exit 1) rather than a silently dropped selection.
-fn resolve_selected_metrics(names: &[String]) -> Option<Vec<big_code_analysis::Metric>> {
-    if names.is_empty() {
-        return None;
-    }
-    crate::metric_alias::validate_diff_metrics(names).unwrap_or_else(|e| die(e));
-    let mut selected: Vec<big_code_analysis::Metric> = names
-        .iter()
-        .filter_map(|name| crate::metric_alias::metric_for_name(name))
-        .collect();
-    selected.sort_unstable();
-    selected.dedup();
-    Some(selected)
-}
-
-fn run_command_metrics(
-    globals: GlobalOpts,
-    args: MetricsArgs,
-    preproc: Option<Arc<PreprocResults>>,
-) {
-    let mut structured = args.structured;
-    let selected_metrics = resolve_selected_metrics(&args.metrics);
-    // `--format text` (issue #604) is a surface alias for the historical
-    // no-`--format` default — the human-readable tree. Collapse it to
-    // `None` here so every downstream guard and the dispatch see the
-    // exact same shape as an omitted flag; the two paths are then
-    // byte-identical by construction.
-    structured.normalize_text_format();
-    let have_format = structured.output_format.is_some();
-    if matches!(structured.output_format, Some(MetricsFormat::Cbor))
-        && structured.output.is_none()
-        && structured.output_dir.is_none()
-    {
-        die(CBOR_STDOUT_ERROR);
-    }
-    let output_target = resolve_structured_output(
-        have_format,
-        structured.output,
-        structured.output_dir,
-        "metrics",
-    );
-    // Build the change-history index once, before the per-file walk, so
-    // the dispatch can attach a `vcs` block to each file (issue #328).
-    // `--vcs` is additive: outside a repo `default_index` warns and
-    // returns `None`, so the AST metrics still emit (vcs block omitted).
-    //
-    // Only the structured-output path renders the `vcs` block; the
-    // human-readable dump (no `--format`) cannot show it. So skip the
-    // (expensive) history walk entirely when no format is selected,
-    // warning that `--vcs` had no effect rather than doing the walk and
-    // silently discarding it.
-    // `--vcs-per-function` (issue #329) implies `--vcs`: the file-level
-    // block is still attached, plus a blame-derived block on every nested
-    // function space. The blame engine is built once here and shared
-    // read-only across workers.
-    let want_vcs = args.vcs || args.vcs_per_function;
-    let (vcs_index, vcs_blame) = match (want_vcs, structured.output_format.is_some()) {
-        (true, true) => {
-            let index = crate::vcs_command::default_index(&globals);
-            let blame = if args.vcs_per_function {
-                crate::vcs_command::default_blame(&globals)
-            } else {
-                None
-            };
-            (index, blame)
-        }
-        (true, false) => {
-            warn(
-                "--vcs / --vcs-per-function has no effect without --format: the \
-                 human-readable metrics view does not render the vcs block",
-            );
-            (None, None)
-        }
-        (false, _) => (None, None),
-    };
-    let pretty = structured.pretty;
-    let fmt = structured.output_format;
-    let action = Action::Metrics {
-        format: fmt,
-        pretty,
-    };
-    // #663 counters: track analyzable output and explicitly-named files
-    // skipped for unrecognized language, so a run that produced nothing
-    // because every named file was unrecognized exits 1.
-    let output_produced = Arc::new(AtomicUsize::new(0));
-    let explicit_unrecognized = Arc::new(AtomicUsize::new(0));
-    // Single-file aggregate mode (#669) streams each space through a
-    // channel, collected and written once after the walk.
-    let (aggregate_tx, aggregate_rx) = match output_target {
-        StructuredOutput::AggregateFile(_) => {
-            let (tx, rx) = std::sync::mpsc::channel();
-            (Some(Mutex::new(tx)), Some(rx))
-        }
-        _ => (None, None),
-    };
-    let cfg = Config {
-        output_dir: match &output_target {
-            StructuredOutput::Dir(dir) => Some(dir.clone()),
-            _ => None,
-        },
-        aggregate_tx,
-        vcs_index,
-        vcs_blame,
-        selected_metrics,
-        output_produced: Some(Arc::clone(&output_produced)),
-        explicit_unrecognized: Some(Arc::clone(&explicit_unrecognized)),
-        ..Config::new(action, &globals, preproc)
-    };
-    run_walk(globals, cfg);
-    if let (StructuredOutput::AggregateFile(out), Some(rx)) = (&output_target, aggregate_rx) {
-        write_aggregate(fmt, rx.into_iter().collect::<Vec<_>>(), out, pretty);
-    }
-    enforce_explicit_unrecognized(&output_produced, &explicit_unrecognized);
-}
-
-/// Exit 1 when a `metrics` / `ops` run produced no analyzable output
-/// *and* skipped at least one explicitly-named file for an unrecognized
-/// language (#663) — the parallel to #596's nonexistent-explicit-path
-/// error. A mixed run that analyzed at least one file still exits 0 (the
-/// per-file warning already fired on stderr).
-fn enforce_explicit_unrecognized(
-    output_produced: &AtomicUsize,
-    explicit_unrecognized: &AtomicUsize,
-) {
-    if output_produced.load(Ordering::Relaxed) == 0
-        && explicit_unrecognized.load(Ordering::Relaxed) > 0
-    {
-        die("no output produced: every explicitly-named file had an \
-             unrecognized language; pass --language to force a parser");
-    }
-}
-
-/// Write the streamed per-file results as ONE aggregate document to `out`
-/// (#669). Reuses the same `--format` the per-file directory mode uses:
-/// CSV concatenates each space's rows; the generic formats emit a
-/// top-level array (TOML under a `files` key). `fmt` is `Some` here — an
-/// aggregate file without a structured format is rejected upstream by
-/// `resolve_structured_output`.
-fn write_aggregate(
-    fmt: Option<MetricsFormat>,
-    items: Vec<AggregateItem>,
-    out: &Path,
-    pretty: bool,
-) {
-    let Some(fmt) = fmt else {
-        return;
-    };
-    // `metrics` streams `FuncSpace`; `ops` streams `Ops`. The two never
-    // mix in one run (separate subcommands), so the first item's variant
-    // determines the element type for the whole aggregate.
-    let is_ops = matches!(items.first(), Some(AggregateItem::Ops(_)));
-    let result = if is_ops {
-        let ops: Vec<Ops> = items
-            .into_iter()
-            .filter_map(|item| match item {
-                AggregateItem::Ops(o) => Some(*o),
-                AggregateItem::Metrics(..) => None,
-            })
-            .collect();
-        match fmt.dispatch() {
-            MetricsDispatch::Generic(g) => crate::formats::dump_aggregate(g, &ops, out, pretty),
-            // CSV is rejected upstream for `ops` (its column schema is
-            // metric-shaped), so this arm is unreachable; fall back to the
-            // generic JSON array to stay exhaustive without a banned
-            // `panic!`.
-            MetricsDispatch::Csv => {
-                crate::formats::dump_aggregate(GenericFormat::Json, &ops, out, pretty)
-            }
-        }
-    } else {
-        let spaces_with_path: Vec<(FuncSpace, PathBuf)> = items
-            .into_iter()
-            .filter_map(|item| match item {
-                AggregateItem::Metrics(space, path) => Some((*space, path)),
-                AggregateItem::Ops(_) => None,
-            })
-            .collect();
-        match fmt.dispatch() {
-            MetricsDispatch::Csv => crate::formats::dump_csv_aggregate(&spaces_with_path, out),
-            MetricsDispatch::Generic(g) => {
-                let spaces: Vec<FuncSpace> = spaces_with_path.into_iter().map(|(s, _)| s).collect();
-                crate::formats::dump_aggregate(g, &spaces, out, pretty)
-            }
-        }
-    };
-    if let Err(e) = result {
-        die(format_args!(
-            "failed to write aggregate output {}: {e}",
-            out.display()
-        ));
-    }
-}
-
-fn run_command_ops(
-    globals: GlobalOpts,
-    args: StructuredArgs,
-    preproc: Option<Arc<PreprocResults>>,
-) {
-    let mut args = args;
-    // `--format text` collapses to the default human-readable tree, exactly
-    // as an omitted `--format` does (issue #604).
-    args.normalize_text_format();
-    let have_format = args.output_format.is_some();
-    if matches!(args.output_format, Some(MetricsFormat::Cbor))
-        && args.output.is_none()
-        && args.output_dir.is_none()
-    {
-        die(CBOR_STDOUT_ERROR);
-    }
-    if let Some(MetricsDispatch::Csv) = args.output_format.map(MetricsFormat::dispatch) {
-        die(
-            "CSV is not supported by `ops` because its column schema is metric-shaped; use `bca metrics --output-format <fmt>`",
-        );
-    }
-    let output_target = resolve_structured_output(have_format, args.output, args.output_dir, "ops");
-    let pretty = args.pretty;
-    let fmt = args.output_format;
-    let action = Action::Ops {
-        format: fmt,
-        pretty,
-    };
-    let output_produced = Arc::new(AtomicUsize::new(0));
-    let explicit_unrecognized = Arc::new(AtomicUsize::new(0));
-    let (aggregate_tx, aggregate_rx) = match output_target {
-        StructuredOutput::AggregateFile(_) => {
-            let (tx, rx) = std::sync::mpsc::channel();
-            (Some(Mutex::new(tx)), Some(rx))
-        }
-        _ => (None, None),
-    };
-    let cfg = Config {
-        output_dir: match &output_target {
-            StructuredOutput::Dir(dir) => Some(dir.clone()),
-            _ => None,
-        },
-        aggregate_tx,
-        output_produced: Some(Arc::clone(&output_produced)),
-        explicit_unrecognized: Some(Arc::clone(&explicit_unrecognized)),
-        ..Config::new(action, &globals, preproc)
-    };
-    run_walk(globals, cfg);
-    if let (StructuredOutput::AggregateFile(out), Some(rx)) = (&output_target, aggregate_rx) {
-        write_aggregate(fmt, rx.into_iter().collect::<Vec<_>>(), out, pretty);
-    }
-    enforce_explicit_unrecognized(&output_produced, &explicit_unrecognized);
-}
-
-fn run_command_report(
-    globals: GlobalOpts,
-    mut args: ReportArgs,
-    manifest: Option<&Manifest>,
-    preproc: Option<Arc<PreprocResults>>,
-) {
-    if let Some(m) = manifest {
-        m.merge_report(&mut args);
-    }
-    let policy = SuppressionPolicy::from_no_suppress(args.no_suppress.unwrap_or(false));
-    if let Some(ref output) = args.output {
-        validate_output_path(output, "report");
-    }
-    let format = args.resolved_format();
-    // Capture the provenance seed-path string before the walk consumes
-    // `args.strip_prefix` into the walker `Config` (issue #680).
-    let prov_paths = report_seed_paths_display(&args);
-    // Build the change-history report (default windows, top = the same
-    // per-table cap) before the AST walk consumes `globals`. `--vcs` is
-    // additive: outside a git tree `build_default_report` warns and
-    // returns `None`, so the report still renders without the section.
-    let mut vcs = args
-        .vcs
-        .then(|| crate::vcs_command::build_default_report(&globals, args.top))
-        .flatten();
-    let (tx, rx) = std::sync::mpsc::channel();
-    // When the change-history section is present, also collect each file's
-    // cyclomatic sum so its hotspot score can be joined after the walk
-    // (issue #615). Only wired for `report --vcs`; a plain `report` leaves
-    // the sender `None` and the hotspot column omitted downstream.
-    let (hotspot_rx, report_hotspot_tx) = if vcs.is_some() {
-        let (htx, hrx) = std::sync::mpsc::channel();
-        (Some(hrx), Some(Mutex::new(htx)))
-    } else {
-        (None, None)
-    };
-    let cfg = Config {
-        markdown_tx: Some(Mutex::new(tx)),
-        report_hotspot_tx,
-        strip_prefix: args.strip_prefix,
-        ..Config::new(Action::Report, &globals, preproc)
-    };
-    run_walk(globals, cfg);
-
-    // ConcurrentRunner::run() consumed Config (and thus the Senders).
-    // All worker threads have joined, so `rx.into_iter()` terminates.
-    let summaries: Vec<FunctionSummary> = rx.into_iter().collect();
-    // Join the per-file hotspot scores onto the change-history rows before
-    // rendering, so `report --vcs` fills the Hotspot column from the AST
-    // metrics computed in this same run (issue #615).
-    if let (Some(default), Some(hotspot_rx)) = (vcs.as_mut(), hotspot_rx) {
-        let cyclomatic_sums: Vec<(PathBuf, f64)> = hotspot_rx.into_iter().collect();
-        default.join_hotspot_scores(&cyclomatic_sums);
-    }
-    let vcs = vcs.map(|d| d.report);
-    let top = args.top;
-
-    // Provenance footer facts (issue #680): version from the package, date
-    // (SOURCE_DATE_EPOCH-overridable) from `provenance`, the user's seed paths
-    // (what they asked to scan, not every walked file), the resolved `--top`,
-    // and whether suppression markers were honored.
-    let date = crate::provenance::resolved_date();
-    let prov = crate::provenance::Provenance {
-        version: env!("CARGO_PKG_VERSION"),
-        date: &date,
-        paths: &prov_paths,
-        top,
-        policy,
-    };
-
-    // Source the report's advisory cutoffs (Actionable Summary, CC note,
-    // Many-Parameters filter) from the manifest `[thresholds]` table when one
-    // is present, so the report's advice matches the configured `bca check`
-    // gate; fall back to the built-in defaults otherwise (issue #630). This
-    // reads the manifest hard-threshold scalars for *presentation* only — it
-    // does not touch the offender-gating path.
-    let advisory = match manifest {
-        Some(m) => AdvisoryThresholds::from_manifest_hard(&m.thresholds().hard),
-        None => AdvisoryThresholds::DEFAULT,
-    };
-
-    // `generate_*_with_vcs` already accept the change-history section as an
-    // `Option`, so dispatch only on the output format and pass the optional
-    // report straight through rather than enumerating the four format×vcs
-    // combinations.
-    let vcs = vcs.as_ref();
-    let report = match format {
-        ReportFormat::Markdown => {
-            generate_report_with_vcs(&summaries, top, policy, &advisory, vcs, Some(&prov))
-        }
-        ReportFormat::Html => {
-            generate_html_report_with_vcs(&summaries, top, policy, &advisory, vcs, Some(&prov))
-        }
-    };
-    write_output_or_stdout(args.output.as_deref(), "write report to", report.as_bytes());
-}
-
-/// The compact seed-path string the provenance footer reports: the `--paths`
-/// the user passed (after manifest merge), joined by `, `, or `.` when none
-/// were given (the implicit current-directory default). Non-UTF-8 path
-/// components fall back to their lossy form for display only — the footer is
-/// human-facing prose, never a map key (cf. the path-as-identifier rule).
-fn report_seed_paths_display(args: &ReportArgs) -> String {
-    let paths = args.seed_paths();
-    if paths.is_empty() {
-        ".".to_owned()
-    } else {
-        paths
-            .iter()
-            .map(|p| p.display().to_string())
-            .collect::<Vec<_>>()
-            .join(", ")
-    }
-}
-
-fn run_command_find(globals: GlobalOpts, args: FindArgs, preproc: Option<Arc<PreprocResults>>) {
-    let line = args.line;
-    let nodes = args.nodes.types;
-    let cfg = Config {
-        line_start: line.line_start,
-        line_end: line.line_end,
-        ..Config::new(Action::Find(nodes.into()), &globals, preproc)
-    };
-    run_walk(globals, cfg);
-}
-
-fn run_command_count(globals: GlobalOpts, args: CountArgs, preproc: Option<Arc<PreprocResults>>) {
-    let collector = CountCollector::new();
-    let cfg = Config {
-        count_lock: Some(collector.clone()),
-        ..Config::new(Action::Count(args.nodes.types.into()), &globals, preproc)
-    };
-    run_walk(globals, cfg);
-
-    // All worker threads have joined, so the collector's `Arc` refcount
-    // is back to one; `into_count` recovers the tally (degrading rather
-    // than panicking if a worker poisoned the inner mutex, issue #445).
-    let count = collector.into_count();
-    println!("{count}");
-}
-
-fn run_command_strip_comments(
-    globals: GlobalOpts,
-    args: StripCommentsArgs,
-    preproc: Option<Arc<PreprocResults>>,
-) {
-    // `--output` is a single-file sink; if more than one input file
-    // matched, every worker would write to the same path (racing,
-    // last-writer-wins, silent data loss). Resolve the file set first
-    // and reject the multi-file case — `--in-place` is the multi-file
-    // verb.
-    let has_output = args.output.is_some();
-    let action = Action::StripComments {
-        in_place: args.in_place,
-        output: args.output,
-    };
-    let mut cfg = Config::new(action, &globals, preproc);
-    let (resolved, num_jobs) = resolve_walk_files(globals);
-    if has_output && resolved.files.len() > 1 {
-        die(format_args!(
-            "--output writes a single file, but {} input files matched; \
-             use --in-place to rewrite multiple files",
-            resolved.files.len()
-        ));
-    }
-    cfg.explicit_seeds = Arc::new(resolved.explicit_files);
-    run_walk_resolved(resolved.files, num_jobs, cfg);
-}
-
-/// Recovers the accumulated [`PreprocResults`] from the shared worker
-/// accumulator once every worker has joined.
-///
-/// Mirrors the panic-free recovery of [`CountCollector::into_count`]
-/// (issue #445): a worker that panicked mid-update poisons the inner
-/// mutex, and a worker that failed to join leaves the `Arc` shared.
-/// Both failure modes degrade to the recovered data rather than
-/// panicking (issue #740). The recovered guard still holds the
-/// fully-applied append-collections, since each worker inserts a
-/// distinct file's entry.
-fn into_preproc_data(preproc_lock: Arc<Mutex<PreprocResults>>) -> PreprocResults {
-    match Arc::try_unwrap(preproc_lock) {
-        Ok(mutex) => mutex
-            .into_inner()
-            .unwrap_or_else(std::sync::PoisonError::into_inner),
-        Err(shared) => {
-            let mut guard = shared
-                .lock()
-                .unwrap_or_else(std::sync::PoisonError::into_inner);
-            // The `Arc` is still shared (a worker failed to join), so we
-            // cannot move the data out by value; take it from behind the
-            // guard, leaving an empty default in its place.
-            std::mem::take(&mut *guard)
-        }
-    }
-}
-
-fn run_command_preproc(globals: GlobalOpts, args: PreprocArgs) {
-    let preproc_lock = Arc::new(Mutex::new(PreprocResults::default()));
-    let output = args.output;
-    let cfg = Config {
-        preproc_lock: Some(preproc_lock.clone()),
-        // PreprocProduce builds its own preproc results; any inbound
-        // `--preproc-data` from globals is intentionally ignored for
-        // this command (the original code passed `None` here too).
-        ..Config::new(Action::PreprocProduce, &globals, None)
-    };
-    let paths = run_walk_collecting(globals, cfg);
-    // Group the analyzed file list by basename for cross-file
-    // `#include` resolution. Computing this from the same resolved
-    // list the workers processed (rather than the library's old
-    // directory-walk callback, which #489 left dead) is what restores
-    // `bca preproc` include resolution — see #495.
-    let all_files = group_files_by_basename(paths);
-
-    let mut data = into_preproc_data(preproc_lock);
-    // Include-resolution diagnostics (self-inclusion, cycles, non-UTF-8
-    // paths, un-preprocessed files) are returned rather than written to
-    // stderr by the library, so the CLI surfaces them here.
-    for diagnostic in fix_includes(&mut data.files, &all_files) {
-        eprintln!("{diagnostic}");
-    }
-
-    let serialized = serde_json::to_string(&data)
-        .unwrap_or_else(|e| die(format_args!("failed to serialize preproc data: {e}")));
-    if let Some(output_path) = output {
-        write_file(&output_path, serialized.as_bytes())
-            .unwrap_or_else(|e| die_io("write preproc output to", &output_path, e));
-    } else {
-        println!("{serialized}");
-    }
-}
-
-/// Canonical contents of a freshly-scaffolded `bca.toml` manifest.
-/// This is the consolidated config `bca` auto-discovers when climbing
-/// from the working directory to the repo root (#374, #483): a bare
-/// `bca check` (no flags) reproduces the gate once this file exists.
-/// The `[thresholds]` values mirror the project's own root `bca.toml`,
-/// so `bca init` hands adopters the same limits the project gates
-/// itself with. The two `[thresholds]` tables are kept in lock-step by
-/// the `init_template_thresholds_match_repo_root` drift test: retuning
-/// the root gate fails that test until this template is updated to
-/// match, and vice versa.
-///
-/// Only the surrounding prose differs by design — the root file
-/// carries the repo-specific rationale for each limit, while this
-/// template keeps the generic editing-rules header. Adopters typically
-/// run `--write-baseline` right after `init`, which pins today's
-/// offenders so subsequent runs fail only on regressions, then tighten
-/// limits over time.
-const INIT_MANIFEST_TEMPLATE: &str = "\
-# bca.toml — project manifest, auto-discovered by `bca` when climbing
-# from the working directory to the repo root. This is the single
-# source of truth for the metric gate: with this file present, a bare
-# `bca check` (no flags) reproduces the full gate, zero-config.
-
-# Scan the whole project, honouring the ignore file below.
-paths = [\".\"]
-exclude_from = \".bcaignore\"
-
-# Rust-only policy: treat `?` (the `try_expression` node) as linear
-# error propagation rather than a branch, so it does not contribute to
-# cyclomatic complexity. Equivalent to `--no-cyclomatic-try` on every
-# invocation (#409). Inert for non-Rust grammars (no other grammar
-# emits the node). Uncomment to opt in if your project is Rust.
-# cyclomatic_count_try = false
-
-# Gate-only options consumed by `bca check` (#599). The baseline pairs
-# with `.bca-baseline.toml` so existing offenders are absorbed; only
-# regressions and new offenders fail the gate. See the Baselines recipe
-# in the book for the bootstrap/refresh/retire flow.
-[check]
-baseline = \".bca-baseline.toml\"
-
-# bca per-function threshold configuration.
-#
-# Editing rules:
-#   * Each key is a stable metric name (or dotted sub-metric name) from
-#     `bca list-metrics` / `bca check --help`. Available names:
-#         cognitive, cyclomatic, cyclomatic.modified,
-#         halstead.volume, halstead.difficulty, halstead.effort,
-#         halstead.time, halstead.bugs,
-#         loc.sloc, loc.ploc, loc.lloc, loc.cloc, loc.blank,
-#         nom, tokens, nexits, nargs,
-#         mi.original, mi.sei, mi.visual_studio,
-#         abc, wmc, npm, npa
-#   * Quote keys containing a dot (TOML requires it).
-#   * Values are per-function limits. A function whose metric exceeds the
-#     limit becomes an offender; the baseline file decides whether it
-#     fails CI.
-#   * Adding a metric is a tightening — regenerate `.bca-baseline.toml`
-#     in the same change so day-one CI does not flip red on offenders
-#     that were previously invisible to the gate.
-#
-# Metrics intentionally NOT gated (yet):
-#   halstead.{volume,difficulty,time,bugs}, cyclomatic.modified, mi.*
-# They are still computed and visible in `bca report markdown|html`.
-[thresholds]
-cognitive = 25
-cyclomatic = 15
-\"halstead.effort\" = 50000
-\"loc.sloc\" = 800
-nom = 30
-nargs = 7
-nexits = 5
-abc = 50
-wmc = 60
-
-# Aggregated-report options for `bca report markdown|html` (#501). By
-# default the report honours in-source suppression markers
-# (`bca: suppress`, `bca: suppress-file`, `#lizard forgives`) and omits a
-# function from a metric's hotspot table when that metric is suppressed
-# for it — matching `bca check` and the SARIF emitter. Uncomment to opt
-# into the raw audit view that lists every offender (equivalent to
-# `bca report --no-suppress`).
-# [report]
-# no_suppress = true
-";
-
-/// Canonical contents of a freshly-scaffolded `.bcaignore`. The
-/// patterns are commented out so the file is a no-op until the
-/// adopter opts in — uncommenting per project. They mirror the
-/// patterns the self-scan workflow walks and the book's adoption
-/// recipe recommends.
-const INIT_BCAIGNORE_TEMPLATE: &str = "\
-# Shared exclude list for `bca --paths . --exclude-from .bcaignore ...`.
-# Patterns are `.gitignore`-style; blank lines and lines whose first
-# non-whitespace character is `#` are skipped.
-#
-# Uncomment any patterns that match generated / vendored / test code
-# you do not want included in the metric gate. The defaults below are
-# the common cases; tailor as needed.
-
-# ./target/**
-# ./dist/**
-# ./build/**
-# ./node_modules/**
-# ./**/*.generated.*
-# ./**/tests/**
-# ./**/*_tests.rs
-";
-
-/// Empty-but-valid baseline file written when the user passes
-/// `--no-baseline`. Matches the shape `baseline::render` would emit
-/// for an empty `BaselineFile`, including the `version` key the
-/// loader requires.
-/// Render the empty-baseline placeholder written by `bca init
-/// --no-baseline`. Built dynamically so the `version` line always
-/// tracks [`baseline::BASELINE_VERSION`] rather than drifting on a
-/// schema bump.
-fn init_empty_baseline_template() -> String {
-    format!(
-        "\
-# bca baseline file. Generated by `bca init --no-baseline`.
-# Populate via:
-#   bca check --write-baseline .bca-baseline.toml
-# A function whose metric value worsens vs. its baselined entry still
-# fails; new offenders also still fail. Refresh with `--write-baseline`
-# when entries become stale.
-version = {}
-",
-        baseline::BASELINE_VERSION
-    )
-}
-
-/// Write the initial `.bca-baseline.toml` for `bca init`: an empty
-/// template when `no_baseline`, otherwise a populated one produced by the
-/// same `bca check --write-baseline` path so it is byte-identical to a
-/// manual bootstrap. Split out of `run_command_init` so scaffolding the
-/// config files and producing the baseline read as the two distinct steps
-/// they are (#969).
-fn scaffold_baseline(
-    globals: GlobalOpts,
-    target: &Path,
-    manifest_path: &Path,
-    baseline_path: &Path,
-    no_baseline: bool,
-    preproc: Option<Arc<PreprocResults>>,
-) {
-    if no_baseline {
-        write_atomic(baseline_path, init_empty_baseline_template().as_bytes())
-            .unwrap_or_else(|e| die_io("write", baseline_path, e));
-        eprintln!(
-            "bca init: wrote empty {} (populate via `bca check --write-baseline {}`)",
-            baseline_path.display(),
-            baseline_path.display(),
-        );
-        return;
-    }
-
-    // Reuse the same code path `bca check --write-baseline` uses so the
-    // produced baseline is byte-identical to one a manual bootstrap would
-    // write. The thresholds we just wrote are loaded from the scaffolded
-    // `bca.toml` to keep this consistent with what the user will use
-    // day-to-day.
-    let check_args = CheckArgs {
-        positional: crate::PositionalPaths::default(),
-        selection: crate::WalkSelectionArgs::default(),
-        tuning: crate::WalkTuningArgs::default(),
-        preproc: crate::PreprocConsumeArgs::default(),
-        thresholds: Vec::new(),
-        config: Some(manifest_path.to_path_buf()),
-        no_fail: false,
-        no_suppress: false,
-        report_suppressed: false,
-        output_format: None,
-        output: None,
-        baseline: None,
-        write_baseline: Some(Some(baseline_path.to_path_buf())),
-        no_summary: true,
-        since: None,
-        changed_only: false,
-        github_annotations: crate::CiDetect::Auto,
-        summary_file: None,
-        no_remediation: true,
-        print_effective_config: None,
-        headroom: None,
-        tier: TierSpec::Hard,
-        exit_codes: None,
-        strict_exit_codes: false,
-        baseline_line_tolerance: None,
-        baseline_fuzzy_match: None,
-        check_exclude: Vec::new(),
-        check_exclude_from: None,
-    };
-    // `run_check` early-exits after `write_check_baseline` runs, so it
-    // returns normally on success here.
-    let mut walk_globals = globals;
-    if walk_globals.paths.is_empty() {
-        walk_globals.paths.push(target.to_path_buf());
-    }
-    // `init` writes its baseline from the manifest it just scaffolded, so
-    // it deliberately bypasses manifest discovery (passing `None`) — the
-    // freshly-written `bca.toml` is the source of truth here, supplied
-    // directly via `--config`.
-    run_check(walk_globals, check_args, None, preproc);
-    eprintln!("bca init: wrote {}", baseline_path.display());
-}
-
-fn run_command_init(globals: GlobalOpts, args: InitArgs, preproc: Option<Arc<PreprocResults>>) {
-    let target = args.dir.clone().unwrap_or_else(|| PathBuf::from("."));
-    if !target.exists() {
-        die(format_args!(
-            "bca init: target directory does not exist: {}",
-            target.display()
-        ));
-    }
-    if !target.is_dir() {
-        die(format_args!(
-            "bca init: target path is not a directory: {}",
-            target.display()
-        ));
-    }
-
-    let manifest_path = target.join("bca.toml");
-    let bcaignore_path = target.join(".bcaignore");
-    let baseline_path = target.join(".bca-baseline.toml");
-
-    // Refuse to clobber existing files unless --force. List every
-    // blocker in one shot so the user can decide whether to delete /
-    // back up before retrying, rather than fixing one and re-running
-    // to discover the next.
-    if !args.force {
-        let list = [&manifest_path, &bcaignore_path, &baseline_path]
-            .into_iter()
-            .filter(|p| p.exists())
-            .map(|p| format!("  {}", p.display()))
-            .collect::<Vec<_>>()
-            .join("\n");
-        if !list.is_empty() {
-            die(format_args!(
-                "bca init: refusing to overwrite existing files (pass --force to clobber):\n{list}"
-            ));
-        }
-    }
-
-    write_atomic(&manifest_path, INIT_MANIFEST_TEMPLATE.as_bytes())
-        .unwrap_or_else(|e| die_io("write", &manifest_path, e));
-    eprintln!("bca init: wrote {}", manifest_path.display());
-
-    write_atomic(&bcaignore_path, INIT_BCAIGNORE_TEMPLATE.as_bytes())
-        .unwrap_or_else(|e| die_io("write", &bcaignore_path, e));
-    eprintln!("bca init: wrote {}", bcaignore_path.display());
-
-    // The config files exist; now produce the initial baseline (empty or
-    // populated from the freshly-scaffolded manifest).
-    scaffold_baseline(
-        globals,
-        &target,
-        &manifest_path,
-        &baseline_path,
-        args.no_baseline,
-        preproc,
-    );
-
-    eprintln!(
-        "bca init: done. Next steps:\n  \
-         1. Review {} and tighten/loosen thresholds for your codebase.\n  \
-         2. Uncomment relevant patterns in {}.\n  \
-         3. Run `bca check` to verify the gate (the manifest is auto-discovered).",
-        manifest_path.display(),
-        bcaignore_path.display(),
-    );
-}
-
-/// Diff two baseline files and print the structured result (issue #382).
-///
-/// Both files are loaded through [`load_baseline`] — the same reader
-/// `bca check` uses — so a supported legacy version is migrated on read
-/// and an unsupported version dies with a clear message (exit 1) rather
-/// than silently no-matching every entry. The matcher's `tolerance` and
-/// `fuzzy` parameters do not influence the flattened entry set, so the
-/// defaults are passed: the diff keys on `(path, qualified, metric)`
-/// regardless.
-///
-/// Exits 0 on success by default — the diff is informational, not a
-/// gate. With `--exit-code`, exits 2 ([`crate::EXIT_GATE_BREACH`]) when
-/// the filtered diff is non-empty (git-diff-style opt-in); a tool error
-/// exits 1 regardless.
-fn run_command_diff_baseline(args: DiffBaselineArgs) {
-    // Validate `--output` before the (cheaper, but still avoidable) load
-    // so a bad path fails fast, mirroring `report` / `exemptions`.
-    if let Some(ref output) = args.output {
-        validate_output_path(output, "diff-baseline");
-    }
-    let old = load_baseline(&args.old, baseline::DEFAULT_LINE_TOLERANCE, false);
-    let new = load_baseline(&args.new, baseline::DEFAULT_LINE_TOLERANCE, false);
-    let diff = BaselineDiff::compute(&old.diff_entries(), &new.diff_entries());
-    let filter = SectionFilter::from_flags([
-        args.added_only,
-        args.removed_only,
-        args.worsened_only,
-        args.improved_only,
-    ]);
-    let rendered = match args.format {
-        OutputFormat::Text => diff.render_tty(filter, &args.strip_prefix),
-        OutputFormat::Markdown => diff.render_markdown(filter, &args.strip_prefix),
-        // Serialization of a fixed-shape struct of owned scalars cannot
-        // fail in practice; surface any future error as a tool error
-        // rather than panicking.
-        OutputFormat::Json => diff
-            .render_json()
-            .unwrap_or_else(|e| die(format_args!("failed to serialize diff to JSON: {e}"))),
-    };
-    write_output_or_stdout(
-        args.output.as_deref(),
-        "write diff-baseline to",
-        rendered.as_bytes(),
-    );
-    // Opt-in metric-gate signal (#692): exit 2 when the filtered diff
-    // carries any entry, so CI can detect "something changed" without
-    // parsing the output. Off by default — the diff stays informational.
-    if args.exit_code && !diff.is_empty_under(filter) {
-        process::exit(crate::EXIT_GATE_BREACH);
-    }
-}
-
-fn run_command_diff(globals: GlobalOpts, args: crate::DiffArgs) {
-    // Validate every `--metric` name against the catalog up front, so a
-    // typo (`--metric cylomatic`) errors with a did-you-mean (exit 1)
-    // instead of silently filtering the diff to nothing (#662). Reuses
-    // the `check --threshold` known-names + suggestion machinery and
-    // accepts the #514 dotted/alias spellings the diff filter handles.
-    crate::metric_alias::validate_diff_metrics(&args.metric).unwrap_or_else(|e| die(e));
-    // Validate `--output` before the (potentially slow) `--since` analysis
-    // walk so a bad path fails fast, mirroring `report` / `exemptions`.
-    if let Some(ref output) = args.output {
-        validate_output_path(output, "diff");
-    }
-    let diff = if let Some(since_ref) = args.since.as_deref() {
-        // `--since` takes at most one positional (the after-side tree),
-        // which clap binds to `old` first. A second positional (`new`)
-        // is ambiguous in this mode, so reject it with a clear message
-        // rather than silently ignoring it.
-        if args.new.is_some() {
-            die(
-                "bca diff --since takes at most one positional (the after-side tree); \
-                 omit it to diff against the working tree",
-            );
-        }
-        compute_since_diff(&globals, &args, since_ref)
-    } else {
-        // File/dir mode: both positionals are required captured metric
-        // sets, so reaching here with `--since` absent means `old`/`new`
-        // came from the positionals; enforce both are present.
-        let Some(old) = args.old.as_deref() else {
-            die("bca diff: provide two metric-output paths (<old> <new>) or use --since <ref>");
-        };
-        let Some(new) = args.new.as_deref() else {
-            die("bca diff: missing <new> metric-output path (or use --since <ref>)");
-        };
-        crate::metric_diff::MetricDiff::compute(old, new, args.min_change, &args.metric)
-    }
-    .unwrap_or_else(|e| die(format_args!("{e}")));
-    let rendered = match args.format {
-        OutputFormat::Text => diff.render_tty(&args.strip_prefix),
-        OutputFormat::Markdown => diff.render_markdown(&args.strip_prefix),
-        // Serialization of a fixed-shape struct of owned scalars cannot
-        // fail in practice; surface any future error as a tool error
-        // rather than panicking.
-        OutputFormat::Json => diff
-            .render_json()
-            .unwrap_or_else(|e| die(format_args!("failed to serialize diff to JSON: {e}"))),
-    };
-    write_output_or_stdout(args.output.as_deref(), "write diff to", rendered.as_bytes());
-    // Opt-in metric-gate signal (#692): exit 2 when any delta survives the
-    // active `--min-change` / `--metric` filtering. Off by default — the
-    // diff stays informational.
-    if args.exit_code && !diff.is_empty() {
-        process::exit(crate::EXIT_GATE_BREACH);
-    }
-}
-
-fn compute_since_diff(
-    globals: &GlobalOpts,
-    args: &crate::DiffArgs,
-    since_ref: &str,
-) -> Result<crate::metric_diff::MetricDiff, crate::metric_diff::DiffError> {
-    // Hard-error early on an unresolvable ref / non-git checkout, before
-    // creating any temp state, so nothing needs cleaning up on this path.
-    diff::validate_since_ref(since_ref).unwrap_or_else(|reason| die(reason));
-
-    // Both sides are rooted at their own tree top (a `git archive`
-    // extraction of `<ref>` for the before side, the repo root for the
-    // after side) and pair on root-relative keys. Selection — `--paths`
-    // and the optional positional scope — must therefore be *relative*:
-    // an absolute path addresses the live filesystem, not the extracted
-    // <ref> tree, so it would walk the current tree for both sides and
-    // yield a silent all-zero diff. Reject any absolute selector with a
-    // clear message rather than mis-pair.
-    let scope = args.old.as_deref();
-    let selectors = || globals.paths.iter().map(PathBuf::as_path).chain(scope);
-    let absolute_selector = selectors().find(|p| p.is_absolute());
-    if let Some(abs) = absolute_selector {
-        die(format_args!(
-            "diff --since: paths must be relative (got {}); an absolute \
-             path cannot address the extracted <ref> tree — scope with a \
-             relative --paths / positional instead",
-            abs.display()
-        ));
-    }
-    // A relative selector that escapes its walk root via `..` is just as
-    // unpairable as an absolute one (#704): on the before side it climbs
-    // out of the `/tmp/…` extraction of `<ref>` (reaching unrelated host
-    // files or nothing), while on the after side it climbs out of the
-    // repo root — so the two sides walk different trees and the diff
-    // silently mis-pairs. `git archive` extractions never contain `..`
-    // path components, so any `..` in a selector cannot address real
-    // tracked files on the before side regardless. Reject it with the
-    // same clear message rather than produce a bogus all-zero / partial
-    // diff.
-    let escaping_selector = selectors().find(|p| escapes_root(p));
-    if let Some(esc) = escaping_selector {
-        die(format_args!(
-            "diff --since: paths must stay within the tree (got {}); a `..` \
-             component escapes the extracted <ref> tree and the working \
-             tree differently, mis-pairing the diff — scope with an in-tree \
-             relative --paths / positional instead",
-            esc.display()
-        ));
-    }
-
-    // `--paths-from` names a file list outside either tree; it cannot be
-    // resolved consistently against both the extracted <ref> tree and
-    // the working tree, so reject it rather than silently ignore the
-    // user's selection (scope a `--since` diff with --paths / globs).
-    if globals.paths_from.is_some() {
-        die(format_args!(
-            "diff --since: --paths-from is not supported; scope the diff \
-             with --paths / --include / --exclude instead"
-        ));
-    }
-
-    // TempDir auto-removes on drop — including every `?` below — so the
-    // "no leftover temp trees, even on error" acceptance holds without
-    // manual teardown.
-    let before_tree = tempfile::TempDir::new().map_err(io_to_diff_error)?;
-    diff::materialize_tree(since_ref, before_tree.path()).unwrap_or_else(|reason| die(reason));
-
-    let before_json = tempfile::TempDir::new().map_err(io_to_diff_error)?;
-    let before = crate::walk_metric_set(
-        before_tree.path(),
-        side_globals(globals, scope),
-        before_json.path(),
-    )?;
-
-    // After side: always the working tree, rooted at the *git repo root*
-    // (not the process CWD) so its root-relative keys line up with the
-    // before side — a `git archive` of the whole ref tree, always rooted
-    // at the repo top. This lets `bca diff --since` run from any
-    // subdirectory. The optional positional is a relative *scope* folded
-    // into both sides by `side_globals` (#497), never an alternate root,
-    // so a subtree positional (`bca diff --since HEAD src`) selects the
-    // same files on each side instead of mis-rooting the after walk.
-    let after_root = diff::git_repo_root().unwrap_or_else(|reason| die(reason));
-    let after_json = tempfile::TempDir::new().map_err(io_to_diff_error)?;
-    let after =
-        crate::walk_metric_set(&after_root, side_globals(globals, scope), after_json.path())?;
-
-    Ok(crate::metric_diff::MetricDiff::from_sets(
-        &before,
-        &after,
-        args.min_change,
-        &args.metric,
-    ))
-}
-
-/// Does relative `path` lexically escape its walk root? Walks the
-/// components tracking depth: a `..` at depth 0 (or that drives depth
-/// below 0 at any point) reaches above the root. `CurDir` and `Normal`
-/// components stay at or below it. Used to reject `bca diff --since`
-/// selectors that would address different trees on the before/after
-/// sides (#704). Absolute paths are rejected earlier, so this only sees
-/// relative ones.
-fn escapes_root(path: &Path) -> bool {
-    let mut depth: i32 = 0;
-    for component in path.components() {
-        match component {
-            Component::ParentDir => {
-                depth -= 1;
-                if depth < 0 {
-                    return true;
-                }
-            }
-            Component::Normal(_) => depth += 1,
-            // `CurDir` is a no-op; `RootDir`/`Prefix` cannot appear in a
-            // relative path (absolute selectors are rejected upstream).
-            Component::CurDir | Component::RootDir | Component::Prefix(_) => {}
-        }
-    }
-    false
-}
-
-fn side_globals(globals: &GlobalOpts, scope: Option<&Path>) -> GlobalOpts {
-    let mut side = globals.clone();
-    // The `--since` positional is a relative path *scope*, not a tree
-    // root: merge it into `--paths` so both sides walk the same subtree
-    // and pair on the same keys (#497). Both sides are rooted at their
-    // own tree top (the `git archive` extraction / the repo root), so a
-    // scope of `src` selects `src/…` on each — never re-roots one side.
-    if let Some(scope) = scope {
-        side.paths.push(scope.to_path_buf());
-    }
-    if side.paths.is_empty() {
-        side.paths = vec![PathBuf::from(".")];
-    }
-    // `--paths-from` is rejected upstream in `compute_since_diff` (it
-    // cannot resolve against both the <ref> tree and the working tree),
-    // so it is already `None` here — selection is via `--paths`/globs.
-    side
-}
-
-/// Adapt a `std::io::Error` raised while creating temp state for the
-/// `--since` walk into the `DiffError` the caller already renders.
-fn io_to_diff_error(source: std::io::Error) -> crate::metric_diff::DiffError {
-    crate::metric_diff::DiffError::Read {
-        path: PathBuf::from("<temp>"),
-        source,
-    }
-}
-
-/// Default baseline file audited by `bca exemptions` when neither
-/// `--baseline` nor `bca.toml`'s `[check] baseline` is set. Matches the
-/// filename `bca init` scaffolds and `bca check --write-baseline`
-/// defaults to.
-const DEFAULT_BASELINE_FILE: &str = ".bca-baseline.toml";
-
-/// Audit everything the `bca check` gate skips in one report (issue
-/// #386): in-source suppression markers, `[check.exclude]` globs, and
-/// `.bca-baseline.toml` entries.
-///
-/// Read-only and always exits 0 on success — the report is a review
-/// surface, not a gate. Each section is opt-out via the mutually
-/// exclusive `--only-*` flags (none set = all three). The baseline
-/// (`bca.toml` top-level `baseline`) and exclude (`[check] exclude`)
-/// inputs default to the same sources `bca check` reads, so the audit
-/// reflects what the gate would skip.
-fn run_command_exemptions(
-    globals: GlobalOpts,
-    mut args: ExemptionsArgs,
-    manifest: Option<&Manifest>,
-    preproc: Option<Arc<PreprocResults>>,
-) {
-    // Merge `bca.toml` `[check]` defaults (baseline path, exclude globs)
-    // under the CLI flags — CLI wins, mirroring `bca check`.
-    if let Some(m) = manifest {
-        m.merge_exemptions(&mut args);
-    }
-
-    // Validate `--output` before the (slower) walk so a bad path fails
-    // fast, mirroring `run_command_report`.
-    if let Some(ref output) = args.output {
-        validate_output_path(output, "exemptions");
-    }
-
-    // No `--*-only` flag selects every section; one selects just that
-    // one (clap enforces mutual exclusivity).
-    let only_any = args.markers_only || args.excludes_only || args.baseline_only;
-    let want_markers = !only_any || args.markers_only;
-    let want_excludes = !only_any || args.excludes_only;
-    let want_baseline = !only_any || args.baseline_only;
-
-    // Resolve the config-driven sections before the walk: it consumes
-    // `globals`, and a missing exclude-from file or unparseable baseline
-    // should error ahead of the slower tree traversal.
-    let excludes = want_excludes.then(|| resolve_exclude_globs(&args));
-    let baseline = want_baseline.then(|| resolve_baseline_section(&args));
-    let markers = want_markers.then(|| collect_marker_rows(globals, preproc));
-
-    let report = ExemptionsReport {
-        markers,
-        excludes,
-        baseline,
-    };
-    let rendered = report
-        .render(args.format, &args.strip_prefix)
-        .unwrap_or_else(|e| die(format_args!("failed to serialize exemptions to JSON: {e}")));
-    write_output_or_stdout(
-        args.output.as_deref(),
-        "write exemptions report to",
-        rendered.as_bytes(),
-    );
-}
-
-/// Run the suppression-marker walk and return the flattened rows sorted
-/// by `(path, line)` for deterministic output. Files arrive in
-/// worker-completion order, so the sort cannot be skipped even though
-/// each file's markers are already line-sorted by the collector.
-fn collect_marker_rows(
-    globals: GlobalOpts,
-    preproc: Option<Arc<PreprocResults>>,
-) -> Vec<MarkerRow> {
-    let (tx, rx) = std::sync::mpsc::channel();
-    let cfg = Config {
-        exemptions_tx: Some(Mutex::new(tx)),
-        ..Config::new(Action::Exemptions, &globals, preproc)
-    };
-    run_walk(globals, cfg);
-    // ConcurrentRunner::run() consumed Config (and thus the Sender).
-    // All worker threads have joined, so `rx.into_iter()` terminates.
-    let mut rows: Vec<MarkerRow> = rx
-        .into_iter()
-        .flat_map(|FileMarkers { path, markers }| {
-            markers.into_iter().map(move |marker| MarkerRow {
-                path: path.clone(),
-                marker,
-            })
-        })
-        .collect();
-    rows.sort_by(|a, b| {
-        a.path
-            .cmp(&b.path)
-            .then_with(|| a.marker.line.cmp(&b.marker.line))
-    });
-    rows
-}
-
-/// Resolve the `[check.exclude]` glob list for display: the
-/// CLI/manifest `check_exclude` values unioned with the lines of
-/// `--check-exclude-from`, in that order.
-fn resolve_exclude_globs(args: &ExemptionsArgs) -> Vec<String> {
-    let mut globs = args.check_exclude.clone();
-    if let Some(from) = args.check_exclude_from.as_deref() {
-        match read_exclude_patterns_from(from, "--check-exclude-from") {
-            Ok(patterns) => globs.extend(patterns),
-            Err(e) => die(e),
-        }
-    }
-    globs
-}
-
-/// Resolve and load the baseline section. An explicit/manifest
-/// `--baseline` path is loaded through the same reader `bca check` uses
-/// (dying on a missing or unparseable file). With no path configured,
-/// the default `.bca-baseline.toml` is audited when present and reported
-/// as empty otherwise — so a zero-config invocation never errors.
-fn resolve_baseline_section(args: &ExemptionsArgs) -> BaselineSection {
-    let path = if let Some(p) = args.baseline.as_deref() {
-        p.to_path_buf()
-    } else {
-        let default = PathBuf::from(DEFAULT_BASELINE_FILE);
-        if !default.exists() {
-            // Zero-config: no baseline present, report an empty section
-            // rather than erroring.
-            return BaselineSection {
-                path: DEFAULT_BASELINE_FILE.to_owned(),
-                entries: Vec::new(),
-            };
-        }
-        default
-    };
-    let loaded = load_baseline(&path, baseline::DEFAULT_LINE_TOLERANCE, false);
-    let mut entries: Vec<BaselineRow> = loaded
-        .diff_entries()
-        .into_iter()
-        .map(BaselineRow::from)
-        .collect();
-    entries.sort_by(|a, b| {
-        (a.path.as_str(), a.start_line, a.metric.as_str()).cmp(&(
-            b.path.as_str(),
-            b.start_line,
-            b.metric.as_str(),
-        ))
-    });
-    BaselineSection {
-        path: path.display().to_string(),
-        entries,
     }
 }
